### 2024-08-22 #7130 @editorialbot
[PRE REVIEW]: wrenfold: Symbolic code generation for robotics
**Submitting author:** @gareth-cross (<a href="http://orcid.org/0009-0008-3110-0078">Gareth Cross</a>)
**Repository:** https://github.com/wrenfold/wrenfold
**Branch with paper.md** (empty if default branch): gareth/joss
**Version:** v0.0.7
