### 2024-08-22 655cd83
Fix SymForce citation
