+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{Martiros-RSS-22,
-    author    = {Martiros, Hayk and The Skydio Autonomy Team},
+    author    = {Hayk Martiros AND Aaron Miller AND Nathan Bucki AND Bradley Solliday AND Ryan Kennedy AND Jack Zhu AND Tung Dang AND Dominic Pattison AND Harrison Zheng AND Teo Tomic AND Peter Henry AND Gareth Cross AND Josiah VanderMey AND Alvin Sun AND Samuel Wang AND Kristen Holtz},
     title     = {{SymForce: Symbolic Computation and Code Generation for Robotics}},
     booktitle = {Proceedings of Robotics: Science and Systems},
     year      = {2022},
