### 2024-08-22 #7130 @gareth-cross
> 👋 @gareth-cross - I believe the compilation of your paper is failing because you are missing an affiliation declaration. Could you take a look at this? Thanks.

Updated with affiliation section, thanks.
