### 2024-08-22 #7130 @boisgera
> @editorialbot invite @boisgera as editor
>
> 👋 @boisgera - can you take this one on as editor?

Yes, sure! You may need to increase my max number of papers from 4 to 5  (if this limit is automatically enforced).
