### 2024-08-23 01a10dd
Expose cmp() function in the python wrapper
