### 2024-08-23 f35024a
number_set is only present on named_variable, unique_variable
