+++ components/core/wf/expressions/variable.h
@@ -45,33 +45,51 @@ class function_argument_variable {
 class unique_variable {
  public:
   // Create a new variable w/ the next index.
-  unique_variable() : index_{next_unique_variable_index()} {}
+  explicit unique_variable(const number_set set)
+      : index_{next_unique_variable_index()}, set_(set) {}
 
-  bool operator<(const unique_variable& other) const noexcept { return index_ < other.index_; }
-  bool operator==(const unique_variable& other) const noexcept { return index_ == other.index_; }
+  constexpr bool operator<(const unique_variable& other) const noexcept {
+    return std::make_tuple(index_, set_) < std::make_tuple(other.index_, other.set_);
+  }
+  constexpr bool operator==(const unique_variable& other) const noexcept {
+    return std::make_tuple(index_, set_) == std::make_tuple(other.index_, other.set_);
+  }
 
   // Retrieve the underlying index.
   constexpr std::size_t index() const noexcept { return index_; }
+
+  // Numeric set the unique variable belongs to.
+  constexpr number_set set() const noexcept { return set_; }
 
  private:
   static std::size_t next_unique_variable_index();
 
   std::size_t index_;
+  number_set set_;
 };
 
 // A variable w/ a user-provided name.
 class named_variable {
  public:
-  explicit named_variable(std::string name) noexcept : name_{std::move(name)} {}
+  explicit named_variable(std::string name, const number_set set) noexcept
+      : name_{std::move(name)}, set_(set) {}
 
-  bool operator<(const named_variable& other) const noexcept { return name_ < other.name_; }
-  bool operator==(const named_variable& other) const noexcept { return name_ == other.name_; }
+  bool operator<(const named_variable& other) const noexcept {
+    return std::make_tuple(name_, set_) < std::make_tuple(other.name_, other.set_);
+  }
+  bool operator==(const named_variable& other) const noexcept {
+    return std::make_tuple(name_, set_) == std::make_tuple(other.name_, other.set_);
+  }
 
   // String name of the variable.
   constexpr const std::string& name() const noexcept { return name_; }
 
+  // Numeric set the variable belongs to.
+  constexpr number_set set() const noexcept { return set_; }
+
  private:
   std::string name_;
+  number_set set_;
 };
 
 // A named variable used in an expression.
@@ -88,16 +106,24 @@ class variable {
 
   // Construct variable from user-provided name.
   variable(std::string name, const number_set set)
-      : identifier_{std::in_place_type_t<named_variable>(), std::move(name)}, set_(set) {}
+      : identifier_{std::in_place_type_t<named_variable>(), std::move(name), set} {}
 
-  variable(identifier_type identifier, const number_set set) noexcept
-      : identifier_(std::move(identifier)), set_(set) {}
+  explicit variable(identifier_type identifier) noexcept : identifier_(std::move(identifier)) {}
 
   // Access the variant of different variable representations.
   constexpr const auto& identifier() const noexcept { return identifier_; }
 
   // The numeric set this variable belongs to.
-  constexpr number_set set() const noexcept { return set_; }
+  constexpr number_set set() const noexcept {
+    struct visitor {
+      constexpr number_set operator()(const named_variable& x) const { return x.set(); }
+      constexpr number_set operator()(const unique_variable& x) const { return x.set(); }
+      constexpr number_set operator()(const function_argument_variable&) const {
+        return number_set::real;
+      }
+    };
+    return std::visit(visitor{}, identifier_);
+  }
 
   // True if this variable is a `unique_variable`.
   constexpr bool is_unique_variable() const noexcept {
@@ -108,49 +134,46 @@ class variable {
   std::string to_string() const;
 
   // Create a function argument expression.
[…]
