### 2024-08-23 ac446b0
number_set is only present on named_variable, unique_variable
