+++ components/core/test_support/wf_test_support/numeric_testing.h
@@ -146,7 +146,7 @@ struct collect_function_input<scalar_expr> {
                   const scalar_type& scalar) const {
     const auto a = static_cast<float_constant::value_type>(arg);
     const bool added = output.add_substitution(
-        variable{function_argument_variable(arg_index, 0, scalar.numeric_type()), number_set::real},
+        variable{function_argument_variable(arg_index, 0, scalar.numeric_type())},
         make_expr<float_constant>(a));
     WF_ASSERT(added);
   }
@@ -167,8 +167,7 @@ struct collect_function_input<type_annotations::static_matrix<Rows, Cols>> {
         const auto a_ij = static_cast<float_constant::value_type>(arg(i, j));
         const bool added = output.add_substitution(
             variable{function_argument_variable(arg_index, element,
-                                                numeric_primitive_type::floating_point),
-                     number_set::real},
+                                                numeric_primitive_type::floating_point)},
             scalar_expr(a_ij));
         WF_ASSERT(added);
       }
