### 2024-08-23 6fa4394
Add compare() function to the python wrapper to enable canonical ordering from python (#259)
