### 2024-08-24 48e4a5e
Add tests for custom types with integer members
