### 2024-08-24 2d3b447
Add tests for custom types with integer members
