+++ components/core/tests/rust_generation_test/src/lib.rs
@@ -512,3 +512,65 @@ fn test_external_function_call_6() {
         epsilon = 1.0e-15
     );
 }
+
+#[test]
+fn test_integer_argument_1() {
+    assert_eq!(2.0, generated::integer_argument_1(1i64, 2.0));
+    assert_eq!(67.2, generated::integer_argument_1(16i64, 4.2));
+}
+
+#[test]
+fn test_integer_struct_member_1() {
+    use types::MixedNumerics;
+    assert_eq!(
+        0.0,
+        generated::integer_struct_member_1(
+            &MixedNumerics {
+                value: 0.5,
+                mode: 0,
+            },
+            2.0
+        )
+    );
+    assert_eq!(
+        0.3f64.sin() * -2.0,
+        generated::integer_struct_member_1(
+            &MixedNumerics {
+                value: 0.3,
+                mode: -2,
+            },
+            2.0
+        )
+    );
+    assert_eq!(
+        -0.25f64.sin() * 9.0,
+        generated::integer_struct_member_1(
+            &MixedNumerics {
+                value: -0.25,
+                mode: 9,
+            },
+            0.0
+        )
+    );
+    // mode == 1
+    assert_eq!(
+        0.3f64.cos(),
+        generated::integer_struct_member_1(
+            &MixedNumerics {
+                value: 0.3,
+                mode: 1,
+            },
+            1.0
+        )
+    );
+    assert_eq!(
+        (0.3f64 * 1.8).cos(),
+        generated::integer_struct_member_1(
+            &MixedNumerics {
+                value: 0.3,
+                mode: 1,
+            },
+            1.8
+        )
+    );
+}
