+++ components/core/tests/test_expressions_codegen.h
@@ -50,7 +50,9 @@ std::string generate_test_expressions(Generator gen) {
                 arg("x"), arg("y"));
   generate_func(gen, code, &external_function_call_6, "external_function_call_6", arg("x"),
                 arg("y"));
-  generate_func(gen, code, &integer_argument_test_1, "integer_argument_test_1", arg("x"), arg("y"));
+  generate_func(gen, code, &integer_argument_1, "integer_argument_1", arg("x"), arg("y"));
+  generate_func(gen, code, &integer_struct_member_1, "integer_struct_member_1", arg("inputs"),
+                arg("scale"));
   return code;
 }
 
