### 2024-08-31 c94ffd5
Rename code_numeric_type --> numeric_primitive_type for clarity (#261)
