### 2024-09-01 33bdba2
Add standalone versions of the sym.subs and sym.distribute functions
