### 2024-09-01 7cf01ba
Add standalone versions of the sym.subs and sym.distribute functions (#262)
