### 2024-09-01 824f274
Get rid of expression_group type, simplify construction of returned values
