+++ components/core/tests/rust_generation_test/src/generated.rs
@@ -65,10 +65,10 @@ where
   let v002: f64 = v.get(1, 0);
   let v000: f64 = v.get(0, 0);
   let v007: f64 = (v000 * v000 + v002 * v002 + v004 * v004).sqrt();
-  let v011: f64 = (1i64) as f64 / v007;
-  D_v.set(0, 0, v000 * v011);
-  D_v.set(0, 1, v002 * v011);
-  D_v.set(0, 2, v004 * v011);
+  let v010: f64 = (1i64) as f64 / v007;
+  D_v.set(0, 0, v000 * v010);
+  D_v.set(0, 1, v002 * v010);
+  D_v.set(0, 2, v004 * v010);
   v007
 }
 
@@ -240,12 +240,12 @@ pub fn atan2_with_derivatives<>(y: f64, x: f64, D_y: &mut f64, D_x: &mut f64) ->
   // negate: 1
   // total: 8
   
-  let v001: f64 = x;
-  let v000: f64 = y;
-  let v009: f64 = (1i64) as f64 / (v000 * v000 + v001 * v001);
-  *D_y = v001 * v009;
-  *D_x = -(v000 * v009);
-  (v000).atan2(v001)
+  let v000: f64 = x;
+  let v001: f64 = y;
+  let v007: f64 = (1i64) as f64 / (v001 * v001 + v000 * v000);
+  *D_y = v000 * v007;
+  *D_x = -(v001 * v007);
+  (v001).atan2(v000)
 }
 
 #[inline]
@@ -339,10 +339,10 @@ where
   let v0007: f64 = w.get(2, 0);
   let v0003: f64 = w.get(0, 0);
   let v0005: f64 = w.get(1, 0);
-  let v0485: f64 = v0007 * v0007;
-  let v0477: f64 = v0003 * v0003;
-  let v0473: f64 = v0005 * v0005;
-  let v0009: f64 = v0473 + v0477 + v0485;
+  let v0487: f64 = v0007 * v0007;
+  let v0479: f64 = v0003 * v0003;
+  let v0475: f64 = v0005 * v0005;
+  let v0009: f64 = v0475 + v0479 + v0487;
   let v0014: f64 = 0.5f64;
   let v0010: f64 = (v0009).sqrt();
   let v0024: f64 = ((1i64) as f64 + v0009 * 0.25f64).sqrt();
@@ -350,8 +350,8 @@ where
   let v0025: f64 = (1i64) as f64 / v0024;
   let v0017: f64 = (v0016).sin();
   let v0013: f64 = (1i64) as f64 / v0010;
-  let v0476: f64 = v0014 * v0025;
-  let v0470: f64 = v0013 * v0017;
+  let v0478: f64 = v0014 * v0025;
+  let v0472: f64 = v0013 * v0017;
   let v0037: f64 = (v0016).cos();
   let v0011: bool = (1e-16f64) < (v0010);
   let v0027: f64;
@@ -359,135 +359,135 @@ where
   let v0038: f64;
   let v0042: f64;
   if v0011 {
-    v0027 = v0005 * v0470;
-    v0034 = v0007 * v0470;
+    v0027 = v0005 * v0472;
+    v0034 = v0007 * v0472;
     v0038 = v0037;
-    v0042 = v0003 * v0470;
+    v0042 = v0003 * v0472;
   } else {
-    v0027 = v0005 * v0476;
-    v0034 = v0007 * v0476;
+    v0027 = v0005 * v0478;
+    v0034 = v0007 * v0478;
     v0038 = v0025;
-    v0042 = v0003 * v0476;
+    v0042 = v0003 * v0478;
   }
   if let Some(R_D_w) = R_D_w {
-    let v0076: f64;
-    let v0083: f64;
-    let v0093: f64;
[…]
