### 2024-09-01 91dccba
Get rid of expression_group type, simplify construction of returned values
