### 2024-09-03 62ddec7
Get rid of flat-packing of scalar_expr values when evaluating custom types
