+++ components/core/wf/compound_expression.h
@@ -46,8 +46,4 @@ std::vector<scalar_expr> create_expression_elements(const compound_expr& provena
 // Create compound expression with type `custom_type_argument` using the provided type and argument.
 compound_expr create_custom_type_argument(const class custom_type& type, std::size_t arg_index);
 
-// Create compound expression with type `custom_type_construction`.
-compound_expr create_custom_type_construction(const custom_type& type,
-                                              std::vector<scalar_expr> expressions);
-
 }  // namespace wf
