### 2024-09-04 ab208ed
Get rid of expression_group type, use any_expression to store function outputs #263
