+++ components/core/wf/code_generation/function_description.cc
@@ -25,6 +25,13 @@ bool is_identical_struct<argument>::operator()(const argument& a, const argument
 function_description::function_description(std::string name) noexcept
     : impl_(std::make_shared<impl>(std::move(name))) {}
 
+const any_expression& function_description::find_output_expressions(const output_key& key) const {
+  const auto it = impl_->output_expressions.find(key);
+  WF_ASSERT(it != impl_->output_expressions.end(), "Key missing: ({}, {})", key.name,
+            string_from_expression_usage(key.usage));
+  return it->second;
+}
+
 std::variant<scalar_expr, matrix_expr, compound_expr> function_description::add_input_argument(
     const std::string_view name, type_variant type) {
   const argument& arg = add_argument(name, std::move(type), argument_direction::input);
@@ -38,25 +45,24 @@ std::variant<scalar_expr, matrix_expr, compound_expr> function_description::add_
 }
 
 void function_description::add_output_argument(const std::string_view name, type_variant type,
-                                               const bool is_optional,
-                                               std::vector<scalar_expr> expressions) {
+                                               const bool is_optional, any_expression expression) {
   add_argument(name, std::move(type),
                is_optional ? argument_direction::optional_output : argument_direction::output);
 
-  output_key key{
-      is_optional ? expression_usage::optional_output_argument : expression_usage::output_argument,
-      name};
-  impl_->output_expressions.emplace_back(std::move(expressions), std::move(key));
+  impl_->output_expressions.emplace(
+      output_key{is_optional ? expression_usage::optional_output_argument
+                             : expression_usage::output_argument,
+                 name},
+      std::move(expression));
 }
 
 // ReSharper disable once CppMemberFunctionMayBeConst
-void function_description::set_return_value(type_variant type,
-                                            std::vector<scalar_expr> expressions) {
+void function_description::set_return_value(type_variant type, any_expression expression) {
   WF_ASSERT(!impl_->return_value_type.has_value(), "Return value on function `{}` already set.",
             name());
   impl_->return_value_type = std::move(type);
-  impl_->output_expressions.emplace_back(std::move(expressions),
-                                         output_key{expression_usage::return_value, ""});
+  impl_->output_expressions.emplace(output_key{expression_usage::return_value, ""},
+                                    std::move(expression));
 }
 
 // ReSharper disable once CppMemberFunctionMayBeConst
