### 2024-09-04 349549c
Fix SymPy conversion of integer function arguments
