### 2024-09-04 c2d9809
Support integer arguments to generated functions (#260)
Adds support for functions accepting integer arguments, and partially addresses #163.
