+++ components/core/tests/cpp_generation_test.cc
@@ -27,6 +27,31 @@ struct Circle {
 
   Eigen::Vector3d to_vector() const { return {center.x, center.y, radius}; }
 };
+
+// An integer that can be implicitly cast to std::int64_t, and nothing else.
+// We use this to make sure our generated code casts correctly when interfacing with
+// `MixedNumerics`.
+class type_safe_int64_t {
+ public:
+  template <typename T, typename = std::enable_if_t<std::is_integral_v<T>>>
+  constexpr type_safe_int64_t(T x) noexcept : x_(static_cast<std::int64_t>(x)) {}  //  NOLINT
+
+  template <typename T, typename = enable_if_same_t<T, std::int64_t>>
+  constexpr operator T() const noexcept {  //  NOLINT
+    return x_;
+  }
+
+  // Allow explicit cast to double.
+  explicit constexpr operator double() const noexcept { return static_cast<double>(x_); }
+
+ private:
+  std::int64_t x_;
+};
+
+struct MixedNumerics {
+  double value;
+  type_safe_int64_t mode;
+};
 }  // namespace wf::numeric
 
 // These functions don't perform amy meaningful task, they just exist so we can call them from
@@ -483,6 +508,43 @@ TEST(CppGenerationTest, TestExternalFunctionCall6) {
                     1.0e-15);
   ASSERT_EIGEN_NEAR(Eigen::Vector2d(623.4375, -14.5),
                     gen::external_function_call_6(-3.5, 4.75).to_vector(), 1.0e-15);
+}
+
+TEST(CppGenerationTest, TestIntegerArgument1) {
+  using return_type = decltype(gen::integer_argument_1(1, 2.0));
+  static_assert(std::is_same_v<return_type, double>);
+  ASSERT_EQ(2.0, gen::integer_argument_1(numeric::type_safe_int64_t(1), 2.0));
+  ASSERT_EQ(67.2, gen::integer_argument_1(numeric::type_safe_int64_t(16), 4.2));
+}
+
+TEST(CppGenerationTest, TestIntegerOutputValues1) {
+  std::int64_t baz;
+  const auto f = gen::integer_output_values_1(numeric::type_safe_int64_t(4), 2.2, baz);
+  static_assert(std::is_same_v<decltype(f), const std::int64_t>);
+  ASSERT_EQ(static_cast<std::int64_t>(4 * 2.2), f);
+  ASSERT_EQ(static_cast<std::int64_t>(static_cast<double>(4) - 2.2), baz);
+}
+
+// Test that generated code correctly casts integers explicitly.
+TEST(CppGenerationTest, TestIntegerStructMember1) {
+  ASSERT_EQ(0, gen::integer_struct_member_1(numeric::MixedNumerics{0.5, 0}, 2.0));
+  ASSERT_EQ(std::sin(0.3) * -2, gen::integer_struct_member_1(numeric::MixedNumerics{0.3, -2}, 2.0));
+  ASSERT_EQ(std::sin(-0.25) * 9,
+            gen::integer_struct_member_1(numeric::MixedNumerics{-0.25, 9}, 0.0));
+  // mode == 1
+  ASSERT_EQ(std::cos(0.3), gen::integer_struct_member_1(numeric::MixedNumerics{0.3, 1}, 1.0));
+  ASSERT_EQ(std::cos(0.3 * 1.8), gen::integer_struct_member_1(numeric::MixedNumerics{0.3, 1}, 1.8));
+}
+
+// Test that we can assign an integer struct member in an output value.
+TEST(CppGenerationTest, TestIntegerStructMember2) {
[…]
