### 2024-09-04 cd828af
Update version number to 0.1.0
