+++ docs/source/quick_start.rst
@@ -14,11 +14,11 @@ version of wrenfold with:
 **Alternatively**, python wheels may also be obtained from the
 `GitHub Releases Page <https://github.com/wrenfold/wrenfold/releases>`_. Select the ``whl`` file
 appropriate to your OS and python version. For example, for python 3.10 on arm64 OSX you would
-download and install ``wrenfold-0.0.7-cp310-cp310-macosx_11_0_arm64.whl``:
+download and install ``wrenfold-0.1.0-cp310-cp310-macosx_11_0_arm64.whl``:
 
 .. code:: bash
 
-   pip install wrenfold-0.0.7-cp310-cp310-macosx_11_0_arm64.whl
+   pip install wrenfold-0.1.0-cp310-cp310-macosx_11_0_arm64.whl
 
 Then open a python REPL and test that wrenfold can be imported:
 
