+++ components/rust/wrenfold-traits/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wrenfold-traits"
-version = "0.0.7"
+version = "0.1.0"
 edition = "2021"
 license = "MIT"
 authors = ["Gareth <gcross.code@icloud.com>"]
