### 2024-09-04 3a0ddc4
Update version number to 0.1.0
