+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-package.version = "0.0.7"
+package.version = "0.1.0"
 package.license = "MIT"
 package.edition = "2021"
 package.authors = ["Gareth <gcross.code@icloud.com>"]
@@ -16,8 +16,8 @@ members = [
 ]
 
 [workspace.dependencies]
-wrenfold-test-utils = { path = "components/rust/wrenfold-test-utils", version = "0.0.7" }
-wrenfold-traits = { path = "components/rust/wrenfold-traits", version = "0.0.7", features = ["nalgebra"] }
+wrenfold-test-utils = { path = "components/rust/wrenfold-test-utils", version = "0.1.0" }
+wrenfold-traits = { path = "components/rust/wrenfold-traits", version = "0.1.0", features = ["nalgebra"] }
 bspline_rust_test = { path = "examples/bspline/bspline_rust_test" }
 
 # External crates used for tests.
