+++ support/conda/wrenfold/meta.yaml
@@ -1,4 +1,4 @@
-{% set version = "0.0.7" %}
+{% set version = "0.1.0" %}
 
 package:
   name: wrenfold
