### 2024-09-05 1715639
Fix __array__ method on MatrixExpr (missing kwargs) (#265)
