### 2024-09-09 #7130 @gareth-cross
@boisgera Per the suggestion in the original post, I scrolled through the list of reviewers and picked a few that might be roughly suitable for this, project based on background:

- [Abdelhak Bougouffa](https://reviewers.joss.theoj.org/reviewers/2305)
- [Azmyin Md. Kamal](https://reviewers.joss.theoj.org/reviewers/4441)
- [Vissarion Fisikopoulos](https://reviewers.joss.theoj.org/reviewers/483)
- [Luca Ferranti](https://reviewers.joss.theoj.org/reviewers/1784)
- [Martin Huber](https://reviewers.joss.theoj.org/reviewers/3026)
- [Maximilian Stahlberg](https://reviewers.joss.theoj.org/reviewers/1964)
