### 2024-09-14 4b1f42e
Simplify binarization algorithm and cleanup code
