+++ components/core/tests/rust_generation_test/src/generated.rs
@@ -339,10 +339,10 @@ where
   let v0007: f64 = w.get(2, 0);
   let v0003: f64 = w.get(0, 0);
   let v0005: f64 = w.get(1, 0);
-  let v0487: f64 = v0007 * v0007;
+  let v0483: f64 = v0007 * v0007;
   let v0479: f64 = v0003 * v0003;
   let v0475: f64 = v0005 * v0005;
-  let v0009: f64 = v0475 + v0479 + v0487;
+  let v0009: f64 = v0475 + v0479 + v0483;
   let v0014: f64 = 0.5f64;
   let v0010: f64 = (v0009).sqrt();
   let v0024: f64 = ((1i64) as f64 + v0009 * 0.25f64).sqrt();
@@ -380,30 +380,29 @@ where
     let v0130: f64;
     let v0138: f64;
     if v0011 {
-      let v0551: f64 = -((1i64) as f64 / (v0010 * v0010 * v0010));
-      let v0528: f64 = ((1i64) as f64 / v0009) * (v0014 * v0037);
-      let v0250: f64 = v0017 * v0551 + v0528;
+      let v0535: f64 = -((1i64) as f64 / (v0010 * v0010 * v0010));
+      let v0513: f64 = ((1i64) as f64 / v0009) * (v0014 * v0037);
+      let v0250: f64 = v0017 * v0535 + v0513;
       let v0481: f64 = v0003 * v0250;
-      let v0529: f64 = (-0.5f64) * v0472;
+      let v0509: f64 = (-0.5f64) * v0472;
       v0077 = v0005 * v0481;
       v0084 = v0007 * v0481;
-      v0094 = v0475 * v0528 + v0017 * (v0013 + v0475 * v0551);
+      v0094 = v0475 * v0513 + v0017 * (v0013 + v0475 * v0535);
       v0100 = v0005 * (v0007 * v0250);
-      v0109 = v0250 * v0487 + v0472;
-      v0116 = v0003 * v0529;
+      v0109 = v0250 * v0483 + v0472;
+      v0116 = v0003 * v0509;
       v0124 = v0250 * v0479 + v0472;
-      v0130 = v0005 * v0529;
-      v0138 = v0007 * v0529;
+      v0130 = v0005 * v0509;
+      v0138 = v0007 * v0509;
     } else {
       let v0075: f64 = (1i64) as f64 / (v0024 * v0024 * v0024);
       let v0473: f64 = (-0.125f64) * v0075;
-      let v0505: f64 = v0003 * v0473;
       let v0114: f64 = -0.25f64;
-      v0077 = v0005 * v0505;
-      v0084 = v0007 * v0505;
+      v0077 = v0473 * (v0003 * v0005);
+      v0084 = v0473 * (v0003 * v0007);
       v0094 = v0473 * v0475 + v0478;
-      v0100 = v0007 * (v0005 * v0473);
-      v0109 = v0473 * v0487 + v0478;
+      v0100 = v0473 * (v0005 * v0007);
+      v0109 = v0473 * v0483 + v0478;
       v0116 = v0114 * (v0003 * v0075);
       v0124 = v0473 * v0479 + v0478;
       v0130 = v0114 * (v0005 * v0075);
@@ -412,82 +411,82 @@ where
     let v0289: f64 = v0038 * v0084;
     let v0361: f64 = v0038 * v0077;
     let v0295: f64 = v0038 * v0100;
[…]
