+++ components/core/wf/code_generation/binarizer.cc
@@ -1,0 +1,162 @@
+// wrenfold symbolic code generator.
+// Copyright (c) 2024 Gareth Cross
+// For license information refer to accompanying LICENSE file.
+#include "wf/code_generation/binarizer.h"
+
+#include <queue>
+#include <type_traits>
+
+WF_BEGIN_THIRD_PARTY_INCLUDES
+#include <absl/container/flat_hash_set.h>
+WF_END_THIRD_PARTY_INCLUDES
+
+#include "wf/utility/scoped_trace.h"
+
+namespace wf {
+
+using value_pair = std::tuple<ir::value_ptr, ir::value_ptr>;
+
+inline auto make_value_pair(ir::value_ptr a, ir::value_ptr b) {
+  if (a->name() < b->name()) {
+    return std::make_tuple(a, b);
+  } else {
+    return std::make_tuple(b, a);
+  }
+}
+
+inline bool value_pair_ordered_by_name(const value_pair& a, const value_pair& b) noexcept {
+  return std::make_tuple(std::get<0>(a)->name(), std::get<1>(a)->name()) <
+         std::make_tuple(std::get<0>(b)->name(), std::get<1>(b)->name());
+}
+
+template <>
+struct hash_struct<value_pair> {
+  constexpr std::size_t operator()(const value_pair& p) const noexcept {
+    return static_cast<std::size_t>(std::get<0>(p)->name()) << 32 |
+           static_cast<std::size_t>(std::get<1>(p)->name());
+  }
+};
+
+// Determine the unique values in sorted container `operands`, and write them to output iterator
+// `output`.
+template <typename Container, typename OutputContainer>
+static void count_operand_multiplicity(const Container& operands, OutputContainer& output) {
+  WF_ASSERT(std::is_sorted(operands.begin(), operands.end(),
+                           [](auto a, auto b) { return a->name() < b->name(); }),
+            "operands: [{}]", fmt::join(operands, ", "));
+
+  output.clear();
+  for (auto it = operands.begin(); it != operands.end();) {
+    const ir::value_ptr val = *it;
+
+    // Find the next unique operand:
+    const auto next =
+        std::find_if(it, operands.end(), [val](const ir::const_value_ptr v) { return v != val; });
+    const std::size_t count = static_cast<std::size_t>(std::distance(it, next));
+
+    // Save the operand and its count:
+    output.emplace_back(val, count);
+    it = next;
+  }
+}
+
+using pair_table =
+    std::unordered_map<value_pair, vector_of_multiplicities, hash_struct<value_pair>>;
+
+// TODO: This method of counting over counts some times of pairs.
+// For example, in the multiplication:
+//  v0 = x*x*x*x*y
+// There are two instances of x*x, and one instance of x*y.
+// We should not double count all the permutations of `x`, but we presently do.
+template <typename Container>
+static void count_operand_pairs(const Container& container, const binarize_type operation,
+                                const std::unordered_set<std::uint32_t>& active_values,
+                                pair_table& table) {
+  table.reserve(container.size());
+
+  std::vector<std::tuple<ir::value_ptr, std::size_t>> operands_with_multiplicities{};
+  for (const ir::value_ptr val : container) {
+    const bool is_suitable_operation = (operation == binarize_type::mul && val->is_op<ir::mul>()) ||
+                                       (operation == binarize_type::add && val->is_op<ir::add>());
+    if (!is_suitable_operation || val->is_unused()) {
+      // We only binarize multiplications and additions. Unused values are discarded later.
+      continue;
+    }
+    if (active_values.count(val->name()) == 0) {
+      // This value was considered on a previous invocation of binarize_operations and does not
+      // need to be recounted.
+      continue;
+    }
+
[…]
