### 2024-09-14 92f7dad
Simplify binarization algorithm and cleanup code (#266)
