+++ components/core/wf/code_generation/control_flow_graph.cc
@@ -8,6 +8,7 @@
 #include <unordered_set>
 #include <vector>
 
+#include "wf/code_generation/binarizer.h"
 #include "wf/code_generation/factorizer.h"
 #include "wf/code_generation/ir_block.h"
 #include "wf/code_generation/ir_control_flow_converter.h"
@@ -180,8 +181,7 @@ void control_flow_graph::apply_simplifications(const optimization_params& p) {
 
   if (p.binarize_operations) {
     for (const auto& block : blocks_) {
-      binarize_operations<ir::mul>(block.get());
-      binarize_operations<ir::add>(block.get());
+      binarize_operations(block.get());
     }
 #ifdef WF_DEBUG
     assert_invariants();
@@ -332,256 +332,74 @@ static void topological_sort_values(const ir::const_block_ptr block,
   WF_ASSERT_EQ(initial_size, operations.size(), "operations: [{}]", fmt::join(operations, ", "));
 }
 
-using value_pair = std::tuple<ir::value_ptr, ir::value_ptr>;
-
-inline auto make_value_pair(ir::value_ptr a, ir::value_ptr b) {
-  if (a->name() < b->name()) {
-    return std::make_tuple(a, b);
-  } else {
-    return std::make_tuple(b, a);
-  }
-}
-
-template <>
-struct hash_struct<value_pair> {
-  constexpr std::size_t operator()(const value_pair& p) const noexcept {
-    return static_cast<std::size_t>(std::get<0>(p)->name()) << 32 |
-           static_cast<std::size_t>(std::get<1>(p)->name());
-  }
-};
-
-// Order pairs by score then by their names.
-// We reverse the order of names to that lower names are taken first when finding the max element.
-struct order_value_pair_by_occurrence_then_name {
-  constexpr bool operator()(const std::pair<value_pair, std::size_t>& a,
-                            const std::pair<value_pair, std::size_t>& b) const noexcept {
-    if (a.second < b.second) {
-      return true;
-    } else if (a.second > b.second) {
-      return false;
-    }
-    return std::make_tuple(std::get<0>(a.first)->name(), std::get<1>(a.first)->name()) >
-           std::make_tuple(std::get<0>(b.first)->name(), std::get<1>(b.first)->name());
-  }
-};
-
-// Determine the unique values in sorted container `operands`, and write them to output iterator
-// `output`.
-template <typename Container, typename OutputContainer>
-void count_operand_multiplicity(const Container& operands, OutputContainer& output) {
-  WF_ASSERT(std::is_sorted(operands.begin(), operands.end(),
-                           [](auto a, auto b) { return a->name() < b->name(); }),
-            "operands: [{}]", fmt::join(operands, ", "));
-
-  output.clear();
-  for (auto it = operands.begin(); it != operands.end();) {
-    const ir::value_ptr val = *it;
-
-    // Find the next unique operand:
-    const auto next =
-        std::find_if(it, operands.end(), [val](const ir::const_value_ptr v) { return v != val; });
-    const std::size_t count = static_cast<std::size_t>(std::distance(it, next));
-
-    // Save the operand and its count:
-    output.emplace_back(val, count);
-    it = next;
-  }
-}
-
-// TODO: This method of counting over counts some times of pairs.
-// For example, in the multiplication:
-//  v0 = x*x*x*x*y
-// There are two instances of x*x, and one instance of x*y.
-// We should not double count all the permutations of `x`, but we presently do.
-template <typename OpType, typename Container>
-static auto count_operand_pairs(
-    const Container& container,
-    std::unordered_map<value_pair, std::size_t, hash_struct<value_pair>>& counts) {
[…]
