### 2024-09-14 bece378
sym.derivative simplifies to zero when argument is not relevant to differentiand
