+++ components/core/wf/expressions/derivative_expression.cc
@@ -4,6 +4,7 @@
 #include "wf/expressions/derivative_expression.h"
 
 #include "wf/constants.h"
+#include "wf/utility_visitors.h"
 
 namespace wf {
 
@@ -15,6 +16,13 @@ scalar_expr derivative::create(scalar_expr function, scalar_expr arg, int order)
   if (!arg.is_type<variable>()) {
     throw type_error("Derivatives can only be taken with respect to variables. Arg = {}",
                      arg.to_string());
+  }
+
+  // Inspect `derivative` to see if it is not a function of its own argument:
+  //  diff(f(y), x) --> 0, since f(y) is not a function of `x`.
+  // TODO: This operation should really memoize as it searches `function`.
+  if (!is_function_of(function, arg)) {
+    return constants::zero;
   }
 
   if (const derivative* d = get_if<const derivative>(function);
