### 2024-09-14 38de6d0
Enable taking derivatives wrt symbolic function invocations
