### 2024-09-16 6f056ce
Enable taking derivatives wrt symbolic function invocations (#268)
