### 2024-09-25 9dfc448
Add conda-forge instructions to the documentation
