+++ docs/source/quick_start.rst
@@ -11,6 +11,12 @@ version of wrenfold with:
 
    pip install wrenfold
 
+wrenfold is also available as a `conda-forge <https://anaconda.org/conda-forge/wrenfold>`_ package:
+
+.. code:: bash
+
+   conda install -c conda-forge wrenfold
+
 **Alternatively**, python wheels may also be obtained from the
 `GitHub Releases Page <https://github.com/wrenfold/wrenfold/releases>`_. Select the ``whl`` file
 appropriate to your OS and python version. For example, for python 3.10 on arm64 OSX you would
@@ -20,9 +26,15 @@ download and install ``wrenfold-0.1.0-cp310-cp310-macosx_11_0_arm64.whl``:
 
    pip install wrenfold-0.1.0-cp310-cp310-macosx_11_0_arm64.whl
 
-Then open a python REPL and test that wrenfold can be imported:
+Validating the install
+----------------------
+
+After installation, open a python REPL and test that wrenfold can be imported:
 
 .. code:: python
+
+   import wrenfold
+   print(wrenfold.__version__)  # prints: 0.1.0
 
    from wrenfold import sym
 
