### 2024-09-25 bc8fc8a
Add conda-forge instructions to the documentation (#269)
