### 2024-09-30 #7130 @boisgera
> Hi @boisgera - any updates? I can trawl through the list again and find some more names if need be.

Hi @gareth-cross, sorry for the delay.

I have not received any answer from my previous batch of invites (in my institution).

I have not tried any names in the editorial bot suggestion list or your own list yet ; I'll start to invite people from theses sources today (starting with your list).

Best regards,

Sébastien
