### 2024-09-30 #7303 @editorialbot
[REVIEW]: wrenfold: Symbolic code generation for robotics
**Submitting author:** @gareth-cross (<a href="http://orcid.org/0009-0008-3110-0078">Gareth Cross</a>)
**Repository:** https://github.com/wrenfold/wrenfold
**Branch with paper.md** (empty if default branch): gareth/joss
**Version:** v0.2.2
