### 2024-10-01 ceb0c9e
Remove some superfluous compiler flags from CMake
