### 2024-10-01 67fd99d
Simplify documentation build (do not run python script automatically)
