### 2024-10-01 48049cc
Update test tolerances (for the benefit of OSX CI)
