### 2024-10-01 537f32c
Upgrade `setup-miniconda` step, switch to miniforge3 (#271)
* Upgrade setup-miniconda step in CI github action

* Update test tolerances (for the benefit of OSX CI)
