+++ .github/workflows/ci.yml
@@ -162,10 +162,10 @@ jobs:
         run: ctest --output-on-failure
 
       - name: Install rust
-        run: rustup toolchain install 1.77
+        run: rustup toolchain install 1.81
 
       - name: Set default rust version
-        run: rustup default 1.77
+        run: rustup default 1.81
 
       - name: Install clippy
         run: rustup component add rustfmt clippy
