### 2024-10-01 8973aec
Cleanup CMake + fix brittleness with docs build (#270)
* Remove some superfluous compiler flags from CMake

* Automatically exclude dependencies

* Remove -Wno-sign-compare flag

* Simplify documentation build (do not run python script automatically)
