+++ docs/CMakeLists.txt
@@ -50,32 +50,26 @@ if(${SPHINX_FOUND})
   if(NOT DEFINED Python_EXECUTABLE)
     message(FATAL_ERROR "The Python executable could not be located.")
   endif()
-  if(NOT DEFINED PYTHON_COMMAND_ENV_VARIABLES)
-    message(
-      FATAL_ERROR "Variable PYTHON_COMMAND_ENV_VARIABLES must be defined.")
+  if(NOT DEFINED WF_PYTHON_ENV_VARIABLES)
+    message(FATAL_ERROR "Variable WF_PYTHON_ENV_VARIABLES must be defined.")
   endif()
 
   # Get a list of python sources in `wf_python`.
   get_python_library_sources(PYTHON_LIB_SOURCES)
 
-  # The python script we use to generate RST files, and the list of output
-  # files:
-  set(DOC_GENERATION_SCRIPT "${CMAKE_CURRENT_SOURCE_DIR}/generate_api_rst.py")
-  set(GENERATED_RST_FILES
-      source/python_api/ast.rst
-      source/python_api/code_generation.rst
-      source/python_api/enumerations.rst
-      source/python_api/exceptions.rst
-      source/python_api/expressions.rst
-      source/python_api/geometry.rst
-      source/python_api/sym.rst
-      source/python_api/sympy_conversion.rst
-      source/python_api/type_annotations.rst
-      source/python_api/type_info.rst)
-
-  set(HANDWRITTEN_RST_FILES
+  set(DOCUMENTATION_INPUT_FILES
       "${CMAKE_SOURCE_DIR}/README.md"
       "${CMAKE_SOURCE_DIR}/LICENSE"
+      source/_static/arch_dark.png
+      source/_static/arch_light.png
+      source/_static/benchmark_plots/ImuIntegration-clang.html
+      source/_static/benchmark_plots/ImuIntegration-gcc.html
+      source/_static/benchmark_plots/QuatInterpolation-clang.html
+      source/_static/benchmark_plots/QuatInterpolation-gcc.html
+      source/_static/benchmark_plots/QuatLocalCoords-clang.html
+      source/_static/benchmark_plots/QuatLocalCoords-gcc.html
+      source/_static/benchmark_plots/RollingShutterCamera-clang.html
+      source/_static/benchmark_plots/RollingShutterCamera-gcc.html
       source/building.rst
       source/conf.py
       source/cpp_api/index.rst
@@ -84,13 +78,23 @@ if(${SPHINX_FOUND})
       source/index.rst
       source/license.rst
       source/performance.rst
+      source/python_api/ast.rst
+      source/python_api/code_generation.rst
+      source/python_api/enumerations.rst
+      source/python_api/exceptions.rst
+      source/python_api/expressions.rst
+      source/python_api/geometry.rst
       source/python_api/index.rst
+      source/python_api/sym.rst
+      source/python_api/sympy_conversion.rst
+      source/python_api/type_annotations.rst
+      source/python_api/type_info.rst
       source/quick_start_files/main.cpp
       source/quick_start_files/quick_start_script.py
-      source/quick_start_files/rosenbrock.h
       source/quick_start_files/rosenbrock_crate/Cargo.toml
       source/quick_start_files/rosenbrock_crate/src/generated.rs
       source/quick_start_files/rosenbrock_crate/src/lib.rs
+      source/quick_start_files/rosenbrock.h
       source/quick_start.rst
[…]

[…]
