### 2024-10-09 7ef0cc5
Add a benchmark for plain_formatter
