+++ components/core/benchmarks/CMakeLists.txt
@@ -15,5 +15,6 @@ add_benchmark(bench_add_mul.cc)
 add_benchmark(bench_code_generation.cc)
 add_benchmark(bench_distribute.cc)
 add_benchmark(bench_factorizer.cc)
+add_benchmark(bench_formatter.cc)
 add_benchmark(bench_jacobian.cc)
 add_benchmark(bench_power.cc)
