### 2024-10-09 71b5673
Run ci action weekly (#274)
