### 2024-10-09 c055864
V1, use memoization. Benchmark numbers: 2783 --> 150, 1.65 --> 2.29
