+++ components/core/wf/plain_formatter.cc
@@ -14,323 +14,365 @@ WF_END_THIRD_PARTY_INCLUDES
 
 namespace wf {
 
-void plain_formatter::operator()(const scalar_expr& x) { visit(x, *this); }
-void plain_formatter::operator()(const matrix_expr& x) { visit(x, *this); }
-void plain_formatter::operator()(const boolean_expr& x) { visit(x, *this); }
+template <typename T, typename>
+std::string_view plain_formatter::operator()(const T& x) {
+  if constexpr (is_any_expression_v<T>) {
+    return std::visit(*this, x);
+  } else {
+    auto& map = cache_.get<T>();
+    if (auto it = map.find(x); it != map.end()) {
+      return std::string_view{*it->second};
+    } else {
+      std::string converted = visit(x, *this);
+      // Putting into unique_ptr is a bit gross, but we need pointer stability.
+      // I tried an alternative here where the strings are stored in an array, and accessed by
+      // index. It produced no meaningful improvement on the benchmark though.
+      const auto [it_inserted, _] =
+          map.emplace(x, std::make_unique<std::string>(std::move(converted)));
+      return std::string_view{*it_inserted->second};
+    }
+  }
+}
 
-void plain_formatter::operator()(const addition& add) {
+std::string plain_formatter::convert(const scalar_expr& expr) { return convert_impl(expr); }
+std::string plain_formatter::convert(const boolean_expr& expr) { return convert_impl(expr); }
+std::string plain_formatter::convert(const matrix_expr& expr) { return convert_impl(expr); }
+std::string plain_formatter::convert(const compound_expr& expr) { return convert_impl(expr); }
+
+template <typename T>
+std::string plain_formatter::convert_impl(const T& expr) {
+  plain_formatter fmt{};
+  fmt(expr);
+  // Take the output by moving it:
+  auto& map = fmt.cache_.get<T>();
+  auto it = map.find(expr);
+  WF_ASSERT(it != map.end(), "Object must exist in the map");
+  return std::move(*it->second);
+}
+
+std::string plain_formatter::operator()(const addition& add) {
   WF_ASSERT_GE(add.size(), 2);
+  std::string output{};
 
   // Sort into canonical order:
   absl::InlinedVector<scalar_expr, 8> terms{add.begin(), add.end()};
   std::sort(terms.begin(), terms.end(), [](const scalar_expr& a, const scalar_expr& b) {
-    const std::pair<scalar_expr, scalar_expr> a_coeff_mul = as_coeff_and_mul(a);
-    const std::pair<scalar_expr, scalar_expr> b_coeff_mul = as_coeff_and_mul(b);
-    return order_by(std::get<1>(a_coeff_mul), std::get<1>(b_coeff_mul))
-        .and_then_by(std::get<0>(a_coeff_mul), std::get<0>(b_coeff_mul))
-        .is_less_than();
+    const auto [a_coeff, a_mul] = as_coeff_and_mul(a);
+    const auto [b_coeff, b_mul] = as_coeff_and_mul(b);
+    // Order first by multiplicand and then by numerical coefficient:
+    return order_by(a_mul, b_mul).and_then_by(a_coeff, b_coeff).is_less_than();
   });
 
   for (std::size_t i = 0; i < terms.size(); ++i) {
     if (const auto [coeff, multiplicand] = as_coeff_and_mul(terms[i]); is_negative_number(coeff)) {
       if (i == 0) {
         // For the first term, just negate it:
-        output_ += "-";
+        output += "-";
       } else {
         // Subsequent terms are written as subtractions:
-        output_ += " - ";
+        output += " - ";
       }
       if (is_negative_one(coeff)) {
         // Don't multiply by negative one:
-        format_precedence(precedence::addition, multiplicand);
[…]
