### 2024-10-09 dfca263
Add better test coverage on formatting of compound expressions
