### 2024-10-09 a8a53e3
Improve performance when printing large expressions (#275)
* Add a benchmark for plain_formatter

* V1, use memoization. Benchmark numbers: 2783 --> 150, 1.65 --> 2.29

* Add better test coverage on formatting of compound expressions
