+++ docs/source/quick_start_files/rosenbrock_crate/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rosenbrock_test"
-version = "0.1.0"
+version = "0.0.1"
 edition = "2021"
 publish = false
 
