### 2024-10-17 41fcd49
Update yapf pre-commit version to v0.40.2
