+++ .pre-commit-config.yaml
@@ -22,8 +22,8 @@ repos:
     hooks:
       - id: cmake-format
 
-  - repo: https://github.com/pre-commit/mirrors-yapf
-    rev: 'v0.32.0'
+  - repo: https://github.com/google/yapf
+    rev: 'v0.40.2'
     hooks:
       - id: yapf
         additional_dependencies: [toml]
