### 2024-10-17 d183608
Update yapf pre-commit version to v0.40.2 (#277)
