### 2024-10-18 374587a
Update version number to 0.1.1 (#276)
* Update version number to 0.1.1

* Update cibuildwheel to v2.21.3
