+++ components/core/tests/rust_generation_test/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust_generation_test"
-version = "0.1.0"
+version = "0.0.1"
 edition = "2021"
 publish = false
 
