### 2024-10-19 191cd89
Add a stack allocator implementation
