+++ components/core/wf/utility/stack_allocator.h
@@ -1,0 +1,174 @@
+// wrenfold symbolic code generator.
+// Copyright (c) 2024 Gareth Cross
+// For license information refer to accompanying LICENSE file.
+#pragma once
+#include <memory>
+#include <type_traits>
+
+#ifdef WF_STACK_ALLOCATOR_INSTRUMENT
+#include <fmt/core.h>
+#endif
+
+// A simple custom stack allocator.
+//
+// While implementing this I referred to:
+// Andrei Alexandrescu's talk: https://www.youtube.com/watch?v=LIb3L4vKZ7U
+// https://blog.andreiavram.ro/cpp-custom-allocators-first-step
+// https://github.dev/FelixPetriconi/AllocatorBuilder
+//
+// For now I don't implement the composable adapters from Andrei's talk, since stack +
+// std::allocator is the only pair I need initially.
+namespace wf {
+
+namespace internal {
+// Round up `n` to the nearest `alignment` bytes.
+constexpr size_t round_to_alignment(const std::size_t alignment, std::size_t n) noexcept {
+  const std::size_t remainder = n % alignment;
+  if (remainder != 0) {
+    n += alignment - remainder;
+  }
+  return n;
+}
+}  // namespace internal
+
+// Store a stack-allocated buffer with the given size + alignment.
+// We have a pointer `p_` that indicates the first space in the stack. When allocations are
+// performed, we check if capacity exists and then move p_ forward. All allocations and
+// deallocations must be FILO ordered.
+template <std::size_t Capacity, std::size_t Alignment>
+class stack_allocator {
+ public:
+  using size_type = std::size_t;
+  static constexpr std::size_t capacity = Capacity;
+
+  constexpr stack_allocator() noexcept : p_(&data_[0]) {}
+
+  // Allocate storage for `n` bytes.
+  [[nodiscard]] constexpr void* allocate(const size_type n) noexcept {
+    if (n == 0) {
+      return nullptr;
+    }
+    const std::size_t n_aligned = internal::round_to_alignment(Alignment, n);
+    if (p_ + n_aligned > data_ + Capacity) {
+      // Insufficient space.
+      return nullptr;
+    }
+    const auto ptr = p_;
+#ifdef WF_STACK_ALLOCATOR_INSTRUMENT
+    fmt::print("stack_allocator<{}, {}>::allocate({} [aligned: {}]) --> {}\n", Capacity, Alignment,
+               n, n_aligned, static_cast<void*>(ptr));
+#endif
+    p_ += n_aligned;
+    return static_cast<void*>(ptr);
+  }
+
+  // Deallocate storage for `n` bytes, but only if `[ptr, n]` was the last block we allocated.
+  // It is assumed that `owns` has already been checked before invoking this.
+  constexpr void deallocate(void* ptr, const size_type n) noexcept {
+    const std::size_t n_aligned = internal::round_to_alignment(Alignment, n);
+    if (static_cast<char*>(ptr) + n_aligned == p_) {
+#ifdef WF_STACK_ALLOCATOR_INSTRUMENT
+      fmt::print("stack_allocator<{}, {}>::deallocate({}, {} [aligned: {}])\n", Capacity, Alignment,
+                 static_cast<void*>(ptr), n, n_aligned);
+#endif
+      p_ = static_cast<char*>(ptr);
+    }
+  }
+
+  // Test if the allocator owns the provided ptr.
+  constexpr bool owns(const void* ptr) const noexcept {
+    return ptr >= data_ && ptr < data_ + Capacity;
+  }
+
+  // True if the provided pointer is owned by this allocator, and is in the used region of the
+  // stack. This is for unit testing.
+  constexpr bool in_live_range(const void* ptr) const noexcept { return ptr >= data_ && ptr < p_; }
+
[…]
