### 2024-10-19 ed123c0
Add a stack allocator implementation
