+++ components/core/tests/stack_allocator_test.cc
@@ -1,0 +1,251 @@
+// wrenfold symbolic code generator.
+// Copyright (c) 2024 Gareth Cross
+// For license information refer to accompanying LICENSE file.
+#include <gtest/gtest.h>
+
+#include <list>
+#include <vector>
+
+#include "wf/utility/stack_allocator.h"
+
+namespace wf {
+
+static_assert(4 == internal::round_to_alignment(4, 1));
+static_assert(4 == internal::round_to_alignment(4, 4));
+static_assert(8 == internal::round_to_alignment(8, 4));
+static_assert(16 == internal::round_to_alignment(16, 1));
+static_assert(32 == internal::round_to_alignment(16, 32));
+static_assert(32 == internal::round_to_alignment(32, 3));
+
+struct block {
+  void* ptr;
+  std::size_t n;
+};
+
+template <std::size_t Capacity, std::size_t Alignment>
+void test_with_alignment() {
+  stack_allocator<Capacity, Alignment> s{};
+  EXPECT_FALSE(s.owns(nullptr));
+  EXPECT_TRUE(s.is_empty());
+
+  static_assert(!std::is_copy_assignable_v<decltype(s)>);
+  static_assert(!std::is_move_assignable_v<decltype(s)>);
+  static_assert(!std::is_copy_constructible_v<decltype(s)>);
+  static_assert(!std::is_move_constructible_v<decltype(s)>);
+
+  constexpr std::size_t max_allocations = Capacity / Alignment;
+  static_assert(max_allocations > 1);
+
+  std::vector<block> pointers;
+  for (std::size_t i = 0; i < max_allocations; ++i) {
+    // Vary the sizes we allocate, but keep them all below `Alignment`.
+    const std::size_t n = std::max<std::size_t>(i % Alignment, 1);
+    auto ptr = s.allocate(n);
+    EXPECT_NE(nullptr, ptr);
+    EXPECT_EQ(0, reinterpret_cast<std::ptrdiff_t>(ptr) % Alignment);
+    EXPECT_EQ(Alignment * (i + 1), s.used_bytes());
+    EXPECT_TRUE(s.owns(ptr) && s.in_live_range(ptr));
+    pointers.push_back(block{ptr, n});
+  }
+
+  // Should now be full:
+  ASSERT_EQ(nullptr, s.allocate(1));
+
+  // Trying to remove one that was not at the back should do nothing:
+  for (std::size_t i = 0; i + 1 < pointers.size(); ++i) {
+    s.deallocate(pointers[i].ptr, pointers[i].n);
+    EXPECT_TRUE(s.in_live_range(pointers[i].ptr));  //  Was not freed.
+  }
+
+  // still full:
+  ASSERT_EQ(nullptr, s.allocate(1));
+
+  // Now free all the blocks:
+  std::reverse(pointers.begin(), pointers.end());
+  for (std::size_t i = 0; i < pointers.size(); ++i) {
+    s.deallocate(pointers[i].ptr, pointers[i].n);
+    // Still owned, but no longer in live range:
+    EXPECT_TRUE(s.owns(pointers[i].ptr));
+    EXPECT_FALSE(s.in_live_range(pointers[i].ptr));
+  }
+
+  ASSERT_TRUE(s.is_empty());
+
+  // Should be empty, so check that we can fill it back up:
+  for (std::size_t i = 0; i < max_allocations; ++i) {
+    auto ptr = s.allocate(1);
+    EXPECT_NE(nullptr, ptr);
+    EXPECT_EQ(0, reinterpret_cast<std::ptrdiff_t>(ptr) % Alignment);
+    EXPECT_TRUE(s.owns(ptr) && s.in_live_range(ptr));
[…]
