### 2024-10-19 0b6c3fc
Replace PMR allocator with stack allocator (enabled on OSX)
