### 2024-10-19 ea0f8c9
Replace PMR allocator with stack allocator (enabled on OSX)
