### 2024-10-27 0f363cd
Update pre-commit and clang-format versions
