+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.3.0
+    rev: v5.0.0
     hooks:
       - id: check-yaml
         exclude: '(^dependencies/)|(support/conda/.*\.yaml)'
@@ -12,7 +12,7 @@ repos:
         exclude: '^dependencies/'
 
   - repo: https://github.com/pre-commit/mirrors-clang-format
-    rev: v15.0.6
+    rev: v19.1.2
     hooks:
       - id: clang-format
         files: ^(components|test|benchmarks)/
