### 2024-10-27 a27bf92
Update scikit-build-core version (#280)
