+++ pyproject.toml
@@ -23,7 +23,8 @@ classifiers = [
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12"
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13"
 ]
 
 [project.urls]
@@ -33,7 +34,7 @@ Issues = "https://github.com/wrenfold/wrenfold/issues"
 
 [build-system]
 requires = [
-    "scikit-build-core==0.9.8",
+    "scikit-build-core==0.10.7",
     "cmake>=3.20",
     "ninja>=1.5",
     # mypy required for stubgen.
