### 2024-10-27 738f4c1
Update pre-commit and clang-format versions (#281)
