### 2024-10-28 a380985
Move supporting functions into utilities file
