### 2024-10-28 3ddc64e
Adapt usage of libfmt for C++20
