### 2024-10-28 04f1598
Other compilation fixes for C++20
