### 2024-10-29 347d233
Enable c++20 (#282)
* Update libfmt to 11.0.2

* Adapt usage of libfmt for C++20

* Other compilation fixes for C++20

* Force yapf to run serially to work around non-determinism
