### 2024-10-29 c468d96
Support taking derivative wrt the derivative of a symbolic function invocation
