### 2024-10-29 b891fe7
Support taking derivative wrt the derivative of a symbolic function invocation (#283)
