### 2024-11-01 2e267fc
Eliminate needless copy of expression when there is no numeric coeff [200~504ms --> 457ms (9.3%)]
