### 2024-11-01 31771e6
Do not allocate duplicate expressions for integers +2 and -2
