### 2024-11-02 33e246e
Add the from_two_operands path [457ms --> 434ms (5%)]
