### 2024-11-02 a4d93ac
Add aditional test cases on as_coeff_and_mul
