### 2024-11-02 13f6174
Improve performance of multiplications (#284)
* Move out of multiplication_parts [523ms --> 504ms (3.8%)]

* Eliminate needless copy of expression when there is no numeric coeff [200~504ms --> 457ms (9.3%)]

* Do not allocate duplicate expressions for integers +2 and -2

* Clean up misplaced import of overloaded_visit.h

* Add the from_two_operands path [457ms --> 434ms (5%)]

* Add aditional test cases on as_coeff_and_mul
