+++ components/core/wf/number_set.cc
@@ -5,7 +5,7 @@
 
 #include "wf/enumerations.h"
 #include "wf/expression_visitor.h"
-#include "wf/expressions/all_expressions.h"
+#include "wf/utility/overloaded_visit.h"
 
 namespace wf {
 
