### 2024-11-02 6fdd6fc
Add option to allow conditionals to remain in ternary form
