+++ components/python/wrenfold/code_generation.py
@@ -186,7 +186,8 @@ def create_function_description(func: T.Callable[..., CodegenFuncInvocationResul
 def generate_function(func: T.Callable[..., CodegenFuncInvocationResult],
                       generator: T.Union[CppGenerator, RustGenerator, BaseGenerator],
                       name: T.Optional[str] = None,
-                      optimization_params: T.Optional[OptimizationParams] = None) -> str:
+                      optimization_params: T.Optional[OptimizationParams] = None,
+                      convert_ternaries: bool = True) -> str:
     """
     Accept a python function that manipulates symbolic mathematical expressions, and convert it
     to code in the language emitted by ``generator``. This is a three-step process:
@@ -205,10 +206,14 @@ def generate_function(func: T.Callable[..., CodegenFuncInvocationResult],
 
     Args:
         func: A python function with type-annotated arguments. See
-          :func:`wrenfold.code_generation.create_function_description` for notes on the expected signature.
+          :func:`wrenfold.code_generation.create_function_description` for notes on the expected
+          signature.
         generator: Instance of a code generator, eg. :class:`wrenfold.code_generation.CppGenerator`.
         name: Name of the function. If unspecified, ``func.__name__`` will be used.
-        optimization_params: Parameters governing simplifications/optimizations applied to the output code.
+        optimization_params: Parameters governing simplifications/optimizations applied to the
+          output code.
+        convert_ternaries: Whether to convert ternary :func:`wrenfold.sym.where` statements to
+          if-else control flow. Defaults to true.
 
     Returns: Generated code.
 
@@ -243,7 +248,8 @@ def generate_function(func: T.Callable[..., CodegenFuncInvocationResult],
         }
     """
     description = create_function_description(func=func, name=name)
-    func_ast = transpile(description, params=optimization_params)
+    func_ast = transpile(
+        description, optimization_params=optimization_params, convert_ternaries=convert_ternaries)
     return generator.generate(definition=func_ast)
 
 
