+++ components/core/tests/rust_generation_test_2/src/generated.rs
@@ -2,7 +2,7 @@
 #![cfg_attr(rustfmt, rustfmt_skip)]
 
 #[inline]
-#[allow(non_snake_case, clippy::unused_unit, clippy::collapsible_else_if, clippy::needless_late_init, unused_variables)]
+#[allow(non_snake_case, clippy::unused_unit, clippy::collapsible_else_if, clippy::needless_late_init, unused_variables, unused_parens)]
 pub fn lookup_and_compute_inner_product<>(vec: &std::vec::Vec<crate::types::StructType>, a: f64, b: f64) -> f64
 {
   // Operation counts:
