+++ components/core/wf/code_generation/python_code_generator.cc
@@ -1,0 +1,450 @@
+// wrenfold symbolic code generator.
+// Copyright (c) 2024 Gareth Cross
+// For license information refer to accompanying LICENSE file.
+#include "wf/code_generation/python_code_generator.h"
+
+#include <span>
+
+#include "wf/code_generation/ast_formatters.h"
+#include "wf/code_generation/ast_visitor.h"
+
+namespace wf {
+
+constexpr std::string_view python_string_from_float_width(
+    python_generator_float_width width) noexcept {
+  switch (width) {
+    case python_generator_float_width::float32: {
+      return "float32";
+    }
+    case python_generator_float_width::float64: {
+      return "float64";
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+python_code_generator::python_code_generator(python_generator_target target,
+                                             python_generator_float_width float_width, int indent)
+    : target_(target), float_width_(float_width), indent_(static_cast<std::size_t>(indent)) {
+  if (indent < 1) {
+    throw wf::invalid_argument_error("Indentation must be >= 1. Provided value: {}", indent);
+  }
+}
+
+constexpr std::string_view python_matrix_type_from_target(python_generator_target target) noexcept {
+  switch (target) {
+    case python_generator_target::numpy: {
+      return "np.ndarray";
+    }
+    case python_generator_target::pytorch: {
+      return "th.Tensor";
+    }
+    case python_generator_target::jax: {
+      return "jnp.ndarray";
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+std::string python_code_generator::operator()(const scalar_type& scalar) const {
+  // Arguments can be passed as scalars, or as single-element arrays/tensors.
+  // So we annotate with either type.
+  // Once we drop Python 3.9 support, this can be a union written like: `X | Y`
+  // See: https://peps.python.org/pep-0604/
+  const auto matrix_alternative = python_matrix_type_from_target(target_);
+  if (target_ == python_generator_target::numpy) {
+    // Numpy treats float32, float64, bool, int64 as specific types.
+    // So the input could be a python primitive, a numpy primitive type, or a single element array.
+    switch (scalar.numeric_type()) {
+      case numeric_primitive_type::boolean: {
+        return fmt::format("T.Union[bool, np.bool, {}]", matrix_alternative);
+      }
+      case numeric_primitive_type::integral: {
+        return fmt::format("T.Union[int, np.int64, {}]", matrix_alternative);
+      }
+      case numeric_primitive_type::floating_point: {
+        return fmt::format("T.Union[float, np.{}, {}]",
+                           python_string_from_float_width(float_width_), matrix_alternative);
+      }
+    }
+  } else {
+    switch (scalar.numeric_type()) {
+      case numeric_primitive_type::boolean: {
+        return fmt::format("T.Union[bool, {}]", matrix_alternative);
+      }
+      case numeric_primitive_type::integral: {
+        return fmt::format("T.Union[int, {}]", matrix_alternative);
+      }
+      case numeric_primitive_type::floating_point: {
+        return fmt::format("T.Union[float, {}]", matrix_alternative);
+      }
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+std::string python_code_generator::operator()(const matrix_type&) const {
+  return std::string{python_matrix_type_from_target(target_)};
+}
+
+std::string python_code_generator::operator()(const custom_type& custom) const {
+  return custom.name();
+}
+
+std::string python_code_generator::operator()(const argument&) const {
+  throw type_error("No default formatter for `{}` is provided.", ast::camel_case_name<argument>());
+}
+
[…]
