+++ components/python/wrenfold/code_generation.py
@@ -14,6 +14,9 @@ from pywrenfold.gen import (
     FunctionDescription,
     OptimizationParams,
     OutputKey,
+    PythonGenerator,
+    PythonGeneratorFloatWidth,
+    PythonGeneratorTarget,
     RustGenerator,
     cse_function_description,
     transpile,
