### 2024-11-05 334083c
Remove unnecessary parenthetical around numeric negative numeric literals (not required)
