### 2024-11-05 b9d8973
Add option to allow conditionals to remain in ternary form (#285)
