+++ components/core/wf/code_generation/cpp_code_generator.h
@@ -80,6 +80,8 @@ class cpp_code_generator {
 
   virtual std::string operator()(const ast::special_constant& x) const;
 
+  virtual std::string operator()(const ast::ternary& x) const;
+
   virtual std::string operator()(const ast::variable_ref& x) const;
 
   // Accept `ast_element`.

[…]
