+++ components/core/tests/rust_generation_test/src/generated.rs
@@ -421,7 +421,7 @@ where
       let v0513: f64 = ((1i64) as f64 / v0009) * (v0014 * v0037);
       let v0250: f64 = v0017 * v0535 + v0513;
       let v0481: f64 = v0003 * v0250;
-      let v0509: f64 = (-0.5f64) * v0472;
+      let v0509: f64 = -0.5f64 * v0472;
       v0077 = v0005 * v0481;
       v0084 = v0007 * v0481;
       v0094 = v0475 * v0513 + v0017 * (v0013 + v0475 * v0535);
@@ -433,7 +433,7 @@ where
       v0138 = v0007 * v0509;
     } else {
       let v0075: f64 = (1i64) as f64 / (v0024 * v0024 * v0024);
-      let v0473: f64 = (-0.125f64) * v0075;
+      let v0473: f64 = -0.125f64 * v0075;
       let v0114: f64 = -0.25f64;
       v0077 = v0473 * (v0003 * v0005);
       v0084 = v0473 * (v0003 * v0007);
@@ -571,9 +571,9 @@ where
     let v0513: f64 = ((1i64) as f64 / v0009) * (v0014 * v0037);
     let v0075: f64 = (1i64) as f64 / (v0024 * v0024 * v0024);
     let v0250: f64 = v0017 * v0535 + v0513;
-    let v0473: f64 = (-0.125f64) * v0075;
+    let v0473: f64 = -0.125f64 * v0075;
     let v0114: f64 = -0.25f64;
-    let v0509: f64 = (-0.5f64) * v0472;
+    let v0509: f64 = -0.5f64 * v0472;
     let v0481: f64 = v0003 * v0250;
     let v0138: f64 = if v0011 { (v0007 * v0509) } else { (v0114 * (v0007 * v0075)) };
     let v0084: f64 = if v0011 { (v0007 * v0481) } else { (v0473 * (v0003 * v0007)) };
