### 2024-11-05 7c98545
Switch from variadic dimaond inheritance to macro in code_formatting_wrapper (#286)
