+++ components/wrapper/pywrenfold/code_formatting_wrapper.cc
@@ -22,67 +22,54 @@ using namespace py::literals;
 
 namespace wf {
 
-// A list of every AST type that is formattable. We need to expose overridable methods for all of
-// these so that the user can customize formatting in python.
-using all_formattable_types =
-    type_list_concatenate_t<ast::ast_element_types,
-                            type_list<ast::function_definition, ast::function_signature,
-                                      custom_type, matrix_type, scalar_type>>;
+// clang-format off
+#define WF_APPLY_MACRO_TO_FORMATTABLE_TYPES(stamp) \
+  stamp(ast::add) \
+  stamp(ast::assign_temporary) \
+  stamp(ast::assign_output_matrix) \
+  stamp(ast::assign_output_scalar) \
+  stamp(ast::assign_output_struct) \
+  stamp(ast::boolean_literal) \
+  stamp(ast::branch) \
+  stamp(ast::call_external_function) \
+  stamp(ast::call_std_function) \
+  stamp(ast::cast) \
+  stamp(ast::comment) \
+  stamp(ast::compare) \
+  stamp(ast::construct_custom_type) \
+  stamp(ast::construct_matrix) \
+  stamp(ast::declaration) \
+  stamp(ast::divide) \
+  stamp(ast::float_literal) \
+  stamp(ast::get_argument) \
+  stamp(ast::get_field) \
+  stamp(ast::get_matrix_element) \
+  stamp(ast::integer_literal) \
+  stamp(ast::multiply) \
+  stamp(ast::negate) \
+  stamp(ast::optional_output_branch) \
+  stamp(ast::parenthetical) \
+  stamp(ast::special_constant) \
+  stamp(ast::return_object) \
+  stamp(ast::ternary) \
+  stamp(ast::variable_ref) \
+  stamp(ast::function_definition) \
+  stamp(ast::function_signature) \
+  stamp(custom_type) \
+  stamp(matrix_type) \
+  stamp(scalar_type)  // clang-format on
 
-// Declare a formatting operator for type that throws. The user will override this in python.
-// ReSharper disable once CppPolymorphicClassWithNonVirtualPublicDestructor
-template <typename T>
-class declare_unimplemented {
- public:
-  virtual std::string operator()(const T&) const {
-    throw type_error("Missing override for type `{}`: format_{}", ast::camel_case_name<T>(),
-                     T::snake_case_name_str);
+// Macro used to declare an override for a formattable type.
+// We invoke maybe_override, which is part of wrapper_generator<T>
+#define WF_CALL_MAYBE_OVERRIDE(T) \
+  std::string operator()(const T& element) const override final { return maybe_override(element); }
+
+// Macro used to declare methods on `base_code_generator`.
+#define WF_DECLARE_UNIMPLEMENTED(T)                                                          \
+  virtual std::string operator()(const T&) const {                                           \
+    throw type_error("Missing override for type `{}`: format_{}", ast::camel_case_name<T>(), \
+                     T::snake_case_name_str);                                                \
   }
-};
-
-// Inherit once from `declare_unimplemented` for every type in a type list
-template <typename T>
-class declare_all_unimplemented;
-
-// ReSharper disable once CppPolymorphicClassWithNonVirtualPublicDestructor
-template <typename... Ts>
-class declare_all_unimplemented<type_list<Ts...>> : public declare_unimplemented<Ts>... {
- public:
-  using declare_unimplemented<Ts>::operator()...;
-};
-
-// define_override implements an override to the `std::string operator(const T&)` method.
-// We don't need a constructor here because we inherit virtually from Base. We use virtual
-// inheritance because define_override is inherited once for every type we can format.
-template <typename Derived, typename Base, typename T>
-class define_override : public virtual Base {
- public:
-  std::string operator()(const T& element) const override final {
[…]
