### 2024-11-05 ec1c8e3
Add generate_python method to code_generation.py
