+++ components/python/wrenfold/code_generation.py
@@ -70,7 +70,7 @@ def create_function_description(func: T.Callable[..., CodegenFuncInvocationResul
 
     Tip:
       The provided callable must be type annotated so that wrenfold can deduce the type of input
-      expressions required to invoke. See :doc:`type_annotations` for built-in types that can be
+      expressions required to invoke it. See :doc:`type_annotations` for built-in types that can be
       used to annotate functions. The function should return either:
 
         * A sequence of :class:`wrenfold.code_generation.ReturnValue` or
@@ -79,7 +79,7 @@ def create_function_description(func: T.Callable[..., CodegenFuncInvocationResul
           be interpreted as a ``ReturnValue``.
 
     Args:
-      func: A python function with type-annotated arguments.
+      func: A symbolic python function with type-annotated arguments.
       name: String name of the function.
 
     Returns:
@@ -195,20 +195,20 @@ def generate_function(func: T.Callable[..., CodegenFuncInvocationResult],
     Accept a python function that manipulates symbolic mathematical expressions, and convert it
     to code in the language emitted by ``generator``. This is a three-step process:
 
-        #. The signature of the provided function is inspected to generate input arguments. Next, it is invoked
-           and the symbolic outputs are captured in a :class:`wrenfold.code_generation.FunctionDescription`.
-        #. This description is converted to AST using :func:`wrenfold.code_generation.transpile`. Duplicate
-           operations are eliminated during this step, and conditionals are converted to control flow.
-        #. Lastly, the AST is passed to the provided generator to emit usable code.
+        #. The signature of the provided function is inspected to generate symbolic input arguments.
+           Next, it is invoked and the symbolic outputs are recorded.
+        #. The expression tree is flattened and optimized by :func:`wrenfold.code_generation.transpile`.
+           Duplicate operations are eliminated during this step, and conditionals are converted to
+           control flow. The simplified output is converted to a syntax tree.
+        #. Lastly, the syntax is passed to the provided ``generator`` to emit usable code.
 
     Tip:
 
-      The steps above can also be performed individually. For instance, you might wish to pass the output
-      AST to multiple generators if many simultaneous target languages are required. See
-      :func:`wrenfold.code_generation.create_function_description` for an example.
+      For examples of the types of functions that wrenfold can generate, see the
+      `wrenfold repo <https://github.com/wrenfold/wrenfold/tree/main/examples>`_.
 
     Args:
-        func: A python function with type-annotated arguments. See
+        func: A symbolic python function with type-annotated arguments. See
           :func:`wrenfold.code_generation.create_function_description` for notes on the expected
           signature.
         generator: Instance of a code generator, eg. :class:`wrenfold.code_generation.CppGenerator`.
@@ -216,9 +216,11 @@ def generate_function(func: T.Callable[..., CodegenFuncInvocationResult],
         optimization_params: Parameters governing simplifications/optimizations applied to the
           output code.
         convert_ternaries: Whether to convert ternary :func:`wrenfold.sym.where` statements to
-          if-else control flow. Defaults to true.
+          if-else control flow. Defaults to true. You likely want to set this to False when
+          targeting python frameworks that need to trace control-flow, for example PyTorch or JAX.
 
-    Returns: Generated code.
+    Returns:
+      * A string of generated code.
 
     Example:
       >>> from wrenfold.type_annotations import FloatScalar
[…]
