### 2024-11-05 083ab3c
Add generate_python method to code_generation.py
