+++ components/wrapper/tests/python_code_generation_test.py
@@ -1,32 +1,559 @@
 """
 Test generation and execution of python code.
 """
+
+import collections
 import typing as T
 import unittest
 
-from test_base import MathTestBase
+import jax
+import jax.numpy as jnp
+import numpy as np
+import torch as th
+
+from wrenfold import code_generation, external_functions, sym
+from wrenfold.geometry import Quaternion
+from wrenfold.type_annotations import (
+    FloatScalar,
+    IntScalar,
+    Matrix3,
+    Vector2,
+    Vector3,
+    Vector4,
+)
+
+from .test_base import MathTestBase
 
 
-class PythonCodeGenerationBase(MathTestBase):
+def create_evaluator(func: T.Callable) -> T.Callable:
+    """
+    Create a python function that evaluates a symbolic function numerically by walking the
+    expression tree and substituting in floats.
 
-    def test_simple_operations(self):
-        """Test a few simple operations."""
+    We could promote this to a library feature, although handling of custom types is somewhat
+    tricky. For now it only handles scalars and vectors.
+    """
+    desc = code_generation.create_function_description(func=func)
+    output_expressions = desc.output_expressions()
+
+    input_args = [a for a in desc.arguments if a.is_input]
+
+    def evaluator(*args):
+        assert len(args) == len(
+            input_args), f"Mismatch in # of input args: {len(args)} != {len(input_args)}"
+        sub_list: T.List[T.Tuple[sym.Expr, sym.Expr]] = []
+        for numeric_arg, arg_desc in zip(args, input_args):
+            sym_arg = arg_desc.create_symbolic_input()
+            if isinstance(sym_arg, sym.Expr):
+                if isinstance(numeric_arg, (float, np.float32, np.float64)):
+                    numeric_arg = sym.float(float(numeric_arg))
+                elif isinstance(numeric_arg, (int, np.int32, np.int64)):
+                    numeric_arg = sym.integer(int(numeric_arg))
+                else:
+                    raise TypeError(
+                        f"Invalid type for numeric arg: value = {numeric_arg}, type = {type(numeric_arg)}"
+                    )
+                sub_list.append((sym_arg, numeric_arg))
+            elif isinstance(sym_arg, sym.MatrixExpr):
+                assert isinstance(numeric_arg, np.ndarray), f"{type(numeric_arg)}"
+                if len(numeric_arg.shape) > 2:
+                    raise TypeError("Numpy array must be 1D or 2D")
+                matrix = sym.matrix(numeric_arg.astype(float))
+                for var, num in zip(sym_arg.to_flat_list(), matrix.to_flat_list()):
+                    sub_list.append((var, num))
+            else:
+                raise TypeError(f"Not supported yet: {arg_desc.type}")
+
+        # Now produce output expressions:
+        assert len(output_expressions) > 0, "Must be at least one output"
+
+        return_value = None
+        output_arg_dict = dict()
+        for key, output_exprs in output_expressions.items():
+            output_exprs = sym.subs(output_exprs, sub_list).eval()
+            assert isinstance(output_exprs,
+                              (float, np.ndarray)), f"Failed to evaluate output: {key}"
+            if not key.name:
+                return_value = output_exprs
+            else:
+                output_arg_dict[key.name] = output_exprs
+
+        if return_value is None:
+            return output_arg_dict
+        elif len(output_arg_dict) == 0:
+            return return_value
+        else:
[…]
