+++ components/python/wrenfold/code_generation.py
@@ -304,12 +304,16 @@ def generate_python(func: T.Callable[..., CodegenFuncInvocationResult],
       :class:`wrenfold.code_generation.OutputArg` will instead be **returned** from the generated
       function in a dict of key-value pairs. The example listing below illustrates this behavior.
 
+      **Additionally**, remember that your target framework may not be able to reason about your
+      custom types. For example, `jax.jit <https://jax.readthedocs.io/en/latest/_autosummary/jax.jit.html>`_
+      only operates on Jax arrays and standard python types (tuples, lists, dict, etc).
+
     Example:
       >>> import numpy as np
       >>> import jax
       >>>
       >>> from wrenfold import code_generation, sym
-      >>> from wrenfold.type_annotations import Vector3, FloatScalar
+      >>> from wrenfold.type_annotations import Vector3
       >>>
       >>> def foo(x: Vector3, y: Vector3):
       >>>     # A simple test function, with one return value and one output argument.
