+++ components/core/wf/code_generation/expression_group.h
@@ -4,9 +4,7 @@
 #pragma once
 #include <string>
 #include <string_view>
-#include <vector>
 
-#include "wf/expression.h"
 #include "wf/utility/hashing.h"
 
 namespace wf {
