+++ pyproject.toml
@@ -74,4 +74,4 @@ test-requires = ["numpy", "sympy"]
 [tool.isort]
 multi_line_output = 3
 include_trailing_comma = true
-known_first_party = ["wrenfold", "pywrenfold"]
+known_first_party = ["wrenfold", "pywrenfold", "test_base"]

[…]
