### 2024-11-09 2d9838d
Python examples and tests use `-m` module invocation
