### 2024-11-09 315ff50
Python examples and tests use `-m` module invocation (#289)
