### 2024-11-09 3fe50ad
Add stop_derivative expression type
