+++ components/core/wf/expressions/all_expressions.h
@@ -16,6 +16,7 @@
 #include "wf/expressions/power.h"
 #include "wf/expressions/relational.h"
 #include "wf/expressions/special_constants.h"
+#include "wf/expressions/stop_derivative.h"
 #include "wf/expressions/substitute_expression.h"
 #include "wf/expressions/unevaluated.h"
 #include "wf/expressions/variable.h"
