### 2024-11-09 03e62c1
Add C99 generation example + an example of Kannala-Brandt camera model
