+++ examples/c_generation/c_generation.c
@@ -1,0 +1,7 @@
+// wrenfold symbolic code generator.
+// Copyright (c) 2024 Gareth Cross
+// For license information refer to accompanying LICENSE file.
+
+// We test this file from C++ in `c_generation_test.cc`, but we still want to validate that the
+// generated code will compile as C. To that end, we include the generated code in this C module.
+#include "generated.h"
