### 2024-11-09 80f392b
Fix test_wheel.py to reflect #289
