### 2024-11-09 cc30390
Add stop_derivative expression type (#290)
