### 2024-11-09 3080a4c
Fix bug where nested matrix in custom type resulted in segfault
