### 2024-11-09 7ba0c26
Fix bug where nested matrix in custom type resulted in segfault
