### 2024-11-09 d618f92
Add C99 generation example + an example of Kannala-Brandt camera model (#291)
* Add C99 generation example + an example of Kannala-Brandt camera model

* Fix test_wheel.py to reflect #289
