+++ examples/c_generation/c_code_generator.py
@@ -1,0 +1,286 @@
+import typing as T
+
+from wrenfold import ast, code_generation, type_info
+from wrenfold.enumerations import (
+    RelationalOperation,
+    StdMathFunction,
+    SymbolicConstant,
+)
+
+
+# noinspection PyMethodMayBeStatic
+class CCodeGenerator(code_generation.BaseGenerator):
+    """
+    A custom code-generator that emits C code. I have endeavored to make this compatible
+    with C99. Not being a C programmer myself, it might have small mistakes.
+
+    By inheriting `BaseGenerator` and implementing the methods below, you can adapt wrenfold
+    to whatever language or API you like.
+    """
+
+    def __init__(self, indent: int = 2):
+        super().__init__()
+        assert indent > 0, f"indent = {indent}"
+        self._indent: str = " " * indent
+
+    def _indent_and_join(self, lines: T.Iterable[str]) -> str:
+        lines_split = []
+        for line in lines:
+            lines_split.extend(line.splitlines())
+        return self._indent + f"\n{self._indent}".join(lines_split)
+
+    @staticmethod
+    def _format_numeric_type(t: type_info.NumericType) -> str:
+        if t == type_info.NumericType.Bool:
+            return "bool"
+        elif t == type_info.NumericType.Integer:
+            return "int64_t"
+        elif t == type_info.NumericType.Float:
+            return "double"
+        else:
+            raise NotImplementedError(f"Unsupported type: {t}")
+
+    def format_scalar_type(self, t: type_info.ScalarType) -> str:
+        return self._format_numeric_type(t.numeric_type)
+
+    def format_custom_type(self, custom: type_info.CustomType) -> str:
+        return custom.name
+
+    def format_add(self, add: ast.Add) -> str:
+        return " + ".join(self.format(x) for x in add.args)
+
+    def format_assign_output_matrix(self, mat: ast.AssignOutputMatrix) -> str:
+        """Call `set_output_value` on every element of the matrix."""
+        mat_type = mat.arg.type
+        assert isinstance(mat_type, type_info.MatrixType)
+        result = []
+        # Optional args are pointers, non-optional are const spans.
+        addr_of = "&" if not mat.arg.is_optional else ""
+        for index, val in enumerate(mat.value.args):
+            row, col = mat_type.compute_indices(idx=index)
+            result.append(
+                f"set_output_value({addr_of}{mat.arg.name}, {row}, {col}, {self.format(val)});")
+        return "\n".join(result)
+
+    def format_assign_output_scalar(self, scalar: ast.AssignOutputScalar) -> str:
+        """Scalar arguments are passed as non-const pointers we assign to."""
+        return f"*{scalar.arg.name} = {self.format(scalar.value)};"
+
+    def format_assign_output_struct(self, struct: ast.AssignOutputStruct) -> str:
+        """Assign to a struct."""
+        return f"*{struct.arg.name} = {self.format(struct.value)};"
+
+    def format_assign_temporary(self, temp: ast.AssignTemporary) -> str:
+        return f"{temp.left} = {self.format(temp.right)};"
+
+    def format_boolean_literal(self, b: ast.BooleanLiteral) -> str:
+        """We assume the user has included stdbool.h so we can use true/false."""
+        return "true" if b.value else "false"
+
+    def format_branch(self, branch: ast.Branch) -> str:
+        result = f"if ({self.format(branch.condition)}) {{\n"
+        result += self._indent_and_join(self.format(x) for x in branch.if_branch)
+        if len(branch.else_branch) > 0:
+            result += "\n} else {\n"
+            result += self._indent_and_join(self.format(x) for x in branch.else_branch)
+        result += "\n}"
+        return result
+
[…]
