+++ examples/shared_expressions.py
@@ -1,0 +1,157 @@
+"""
+Symbolic expressions that are shared among the provided examples.
+"""
+
+from wrenfold import code_generation, sym, type_annotations
+
+
+def kb_fisheye_distortion(theta: type_annotations.FloatScalar,
+                          coeffs: type_annotations.Vector4) -> sym.Expr:
+    """Evaluate the Kannala-Brandt fisheye distortion curve (theta -> radius)."""
+    k1, k2, k3, k4 = coeffs
+    radius = theta * (1 + k1 * theta ** 2 + k2 * theta ** 4 + k3 * theta ** 6 + k4 * theta ** 8)
+    return radius
+
+
+def kb_fisheye_invert_distortion(
+    radius: type_annotations.FloatScalar,
+    coeffs: type_annotations.Vector4,
+    num_iters: int = 6,
+) -> sym.Expr:
+    """
+    Invert `kb_fisheye_distortion` with newton solver.
+
+    Args:
+        radius: radius in the image plane.
+        coeffs: Radial distortion coefficients [k1, k2, k3, 4].
+        num_iters: Number of iterations to unroll.
+
+    Tip:
+        We symbolically unroll `num_iters` of iteration. You may wish to raise this number if the
+        accuracy proves insufficient for your camera model.
+    """
+    assert num_iters > 0, f"num_iters = {num_iters}"
+
+    theta = 0
+    for iteration in range(0, num_iters):
+        # Evaluate the forward projection model:
+        theta_sym = sym.symbols("theta", nonnegative=True)
+        r_predicted = kb_fisheye_distortion(theta=theta_sym, coeffs=coeffs)
+
+        # Compute derivative wrt theta.
+        r_D_theta = r_predicted.diff(theta_sym).subs(theta_sym, theta)
+
+        error = r_predicted.subs(theta_sym, theta) - radius
+        updated_theta = sym.max(theta - error / r_D_theta, 0)
+
+        if iteration + 1 < num_iters:
+            theta = sym.stop_derivative(updated_theta)
+        else:
+            # Only allow derivative to propagate on the final iteration.
+            # In other words, we will linearize about the _solution_ of the solver.
+            theta = updated_theta
+
+    return theta
+
+
+def kb_camera_projection(
+    p_cam: type_annotations.Vector3,
+    K: type_annotations.Vector4,
+    coeffs: type_annotations.Vector4,
+):
+    """
+    Evaluate the projection model of a Kannala-Brandt camera model. Only the 4 radial distortion
+    coefficients are implemented in this example.
+
+    Args:
+        p_cam: Incident vector in camera coordinates. We assume it has non-zero norm.
+        K: Intrinsic parameters as a 4-element vector [fx, fy, cx, cy].
+        coeffs: Radial distortion coefficients [k1, k2, k3, 4].
+    """
+    # Angle with respect to the optical axis.
+    # We could use acos() here and probably be a bit faster, although we would additionally need to
+    # normalized `p_cam` as well.
+    xy_norm = p_cam[0:2].norm()
+    theta = sym.where(xy_norm > 0, sym.atan2(xy_norm, p_cam[2]), 0)
+
+    # Angle in the image plane:
+    phi = sym.where(xy_norm > 0, sym.atan2(p_cam[1], p_cam[0]), 0)
+
+    # Distort theta with the radial model:
+    r = kb_fisheye_distortion(theta=theta, coeffs=coeffs)
+
[…]
