### 2024-11-09 d846ecc
Fix bug where nested matrix in custom type resulted in segfault (#292)
