### 2024-11-10 25ee3cd
Add note to README of python_generation example
