+++ examples/python_generation/python_generation.py
@@ -1,5 +1,9 @@
 """
 An example of a custom code-generator implemented entirely in python.
+
+NOTE: This example is obsolete because python generation is now built-into the
+wrenfold framework. It will be removed in short order. See `c_generation` for
+an example of adding a new language.
 """
 import dataclasses
 import typing as T
