+++ examples/python_generation/README.md
@@ -1,1 +1,5 @@
 This example demonstrates how to add support for a new target language.
+
+NOTE: This example is obsolete because python generation is now built-into the
+wrenfold framework. It will be removed in short order. See `c_generation` for
+an example of adding a new language.
