### 2024-11-10 b7785d8
Use pip+requirements.txt to install packages on CI
