+++ docs/source/building.rst
@@ -46,29 +46,34 @@ To run python tests you will additionally need:
   * numpy
   * SymPy
 
-To build and run Rust tests you will need:
-
-  * The rust compiler toolchain
-  * On linux and windows: ``pkg-config`` and ``openblas``.
-
 To build documentation:
 
   * `sphinx <https://www.sphinx-doc.org/>`_
   * `furo <https://github.com/pradyunsg/furo>`_
-  * `doxygen <https://www.doxygen.nl>`_
+  * `doxygen <https://www.doxygen.nl/download.html>`_
   * `breathe <https://breathe.readthedocs.io/en/latest/>`_
   * `myst <https://myst-parser.readthedocs.io/>`_
 
-The repo includes conda configuration files that can be used to create a conda environment suitable for building and testing:
+You can use the `requirements.txt <https://github.com/wrenfold/wrenfold/blob/main/requirements.txt>`_
+file to install most of the dependencies (except for doxygen):
 
 .. code:: bash
 
-    conda create -n wf
-    conda env update -n wf --file environment.yml
-    conda env update -n wf --file <PLATFORM>-environment.yml
-    conda activate wf
+  pip install -r requirements.txt
 
-Where ``<PLATFORM>`` is one of ``mac``, ``linux``, or ``windows``. The rust tools must be installed separately with `rustup <https://rustup.rs>`_.
+.. tip::
+
+  The ``requirements.txt`` file is generated by ``pip-compile``:
+
+  .. code:: bash
+
+    pip-compile requirements.in
+
+To build and run Rust tests you will need:
+
+  * The rust compiler toolchain, installed via `rustup <https://rustup.rs>`_
+  * On linux: ``pkg-config`` and ``openblas``.
+
 
 Compilation
 ^^^^^^^^^^^
