+++ windows-environment.yml
@@ -1,7 +1,0 @@
-# This file specifies additional dependencies required to build on windows.
-channels:
-  - conda-forge
-
-dependencies:
-  - pkg-config
-  - openblas
