### 2024-11-10 94ce8c3
Use pip+requirements.txt to install packages on CI (#293)
