+++ requirements.txt
@@ -1,0 +1,118 @@
+#
+# This file is autogenerated by pip-compile with Python 3.10
+# by the following command:
+#
+#    pip-compile requirements.in
+#
+alabaster==0.7.16
+    # via sphinx
+babel==2.16.0
+    # via sphinx
+beautifulsoup4==4.12.3
+    # via furo
+breathe==4.35.0
+    # via -r requirements.in
+certifi==2024.8.30
+    # via requests
+cfgv==3.4.0
+    # via pre-commit
+charset-normalizer==3.4.0
+    # via requests
+cmake==3.30.5
+    # via -r requirements.in
+distlib==0.3.9
+    # via virtualenv
+docutils==0.19
+    # via
+    #   breathe
+    #   myst-parser
+    #   sphinx
+filelock==3.16.1
+    # via virtualenv
+furo==2024.8.6
+    # via -r requirements.in
+identify==2.6.2
+    # via pre-commit
+idna==3.10
+    # via requests
+imagesize==1.4.1
+    # via sphinx
+jinja2==3.1.4
+    # via
+    #   myst-parser
+    #   sphinx
+markdown-it-py==3.0.0
+    # via
+    #   mdit-py-plugins
+    #   myst-parser
+markupsafe==3.0.2
+    # via jinja2
+mdit-py-plugins==0.4.2
+    # via myst-parser
+mdurl==0.1.2
+    # via markdown-it-py
+mpmath==1.3.0
+    # via sympy
+mypy==1.9.0
+    # via -r requirements.in
+mypy-extensions==1.0.0
+    # via mypy
+myst-parser==3.0.1
+    # via -r requirements.in
+ninja==1.11.1.1
+    # via -r requirements.in
+nodeenv==1.9.1
+    # via pre-commit
+numpy==2.1.3
+    # via -r requirements.in
+packaging==24.2
+    # via sphinx
+platformdirs==4.3.6
+    # via virtualenv
+pre-commit==4.0.1
+    # via -r requirements.in
+pygments==2.18.0
+    # via
+    #   furo
+    #   sphinx
+pyyaml==6.0.2
+    # via
+    #   myst-parser
+    #   pre-commit
+requests==2.32.3
+    # via sphinx
+snowballstemmer==2.2.0
+    # via sphinx
+soupsieve==2.6
+    # via beautifulsoup4
+sphinx==6.2.1
+    # via
+    #   -r requirements.in
+    #   breathe
+    #   furo
+    #   myst-parser
+    #   sphinx-basic-ng
+sphinx-basic-ng==1.0.0b2
+    # via furo
+sphinxcontrib-applehelp==2.0.0
+    # via sphinx
+sphinxcontrib-devhelp==2.0.0
+    # via sphinx
+sphinxcontrib-htmlhelp==2.1.0
+    # via sphinx
+sphinxcontrib-jsmath==1.0.1
+    # via sphinx
+sphinxcontrib-qthelp==2.0.0
+    # via sphinx
+sphinxcontrib-serializinghtml==2.0.0
+    # via sphinx
+sympy==1.13.3
+    # via -r requirements.in
+tomli==2.0.2
+    # via mypy
+typing-extensions==4.12.2
+    # via mypy
+urllib3==2.2.3
+    # via requests
+virtualenv==20.27.1
+    # via pre-commit
