+++ environment.yml
@@ -1,23 +1,0 @@
-name: build-env
-
-channels:
-  - conda-forge
-
-dependencies:
-  - python=3.9
-  - cmake>=3.20
-  - ninja>=1.5
-  - mypy
-  # Linting/formatting
-  - pre-commit
-  # Tests:
-  - numpy
-  - sympy
-  # Documentation
-  # We are stuck on an older sphinx until breathe resolves compatibility issues.
-  # Such as: https://github.com/sphinx-doc/sphinx/issues/11490
-  - sphinx==6.2.1
-  - furo
-  - doxygen
-  - breathe==4.35.0
-  - myst-parser
