+++ components/wrapper/CMakeLists.txt
@@ -4,7 +4,11 @@ add_custom_target(wf_stubs ALL)
 # Find stubgen:
 if(NOT DEFINED STUBGEN_PATH)
   message(STATUS "Looking for stubgen...")
-  find_program(STUBGEN_PATH stubgen)
+  if(WF_STUBS_REQUIRED)
+    find_program(STUBGEN_PATH stubgen REQUIRED)
+  else()
+    find_program(STUBGEN_PATH stubgen)
+  endif()
   if(DEFINED STUBGEN_PATH AND NOT STUBGEN_PATH STREQUAL "STUBGEN_PATH-NOTFOUND")
     message(STATUS "stubgen path: ${STUBGEN_PATH}")
   endif()

[…]
