+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.10
 # by the following command:
 #
-#    pip-compile requirements.in
+#    pip-compile --output-file=requirements.txt support/requirements.in support/testing-requirements.in
 #
 alabaster==0.7.16
     # via sphinx
@@ -11,15 +11,15 @@ babel==2.16.0
 beautifulsoup4==4.12.3
     # via furo
 breathe==4.35.0
-    # via -r requirements.in
+    # via -r support/requirements.in
 certifi==2024.8.30
     # via requests
 cfgv==3.4.0
     # via pre-commit
 charset-normalizer==3.4.0
     # via requests
-cmake==3.30.5
-    # via -r requirements.in
+cmake==3.31.0.1
+    # via -r support/requirements.in
 distlib==0.3.9
     # via virtualenv
 docutils==0.19
@@ -30,13 +30,17 @@ docutils==0.19
 filelock==3.16.1
     # via virtualenv
 furo==2024.8.6
-    # via -r requirements.in
+    # via -r support/requirements.in
 identify==2.6.2
     # via pre-commit
 idna==3.10
     # via requests
 imagesize==1.4.1
     # via sphinx
+jax==0.4.35
+    # via -r support/testing-requirements.in
+jaxlib==0.4.35
+    # via jax
 jinja2==3.1.4
     # via
     #   myst-parser
@@ -51,26 +55,37 @@ mdit-py-plugins==0.4.2
     # via myst-parser
 mdurl==0.1.2
     # via markdown-it-py
+ml-dtypes==0.5.0
+    # via
+    #   jax
+    #   jaxlib
 mpmath==1.3.0
     # via sympy
 mypy==1.9.0
-    # via -r requirements.in
+    # via -r support/requirements.in
 mypy-extensions==1.0.0
     # via mypy
 myst-parser==3.0.1
-    # via -r requirements.in
+    # via -r support/requirements.in
 ninja==1.11.1.1
-    # via -r requirements.in
+    # via -r support/requirements.in
 nodeenv==1.9.1
     # via pre-commit
 numpy==2.1.3
-    # via -r requirements.in
+    # via
+    #   -r support/testing-requirements.in
+    #   jax
+    #   jaxlib
+    #   ml-dtypes
+    #   scipy
+opt-einsum==3.4.0
+    # via jax
 packaging==24.2
     # via sphinx
 platformdirs==4.3.6
     # via virtualenv
 pre-commit==4.0.1
-    # via -r requirements.in
+    # via -r support/requirements.in
 pygments==2.18.0
     # via
     #   furo
@@ -81,13 +96,17 @@ pyyaml==6.0.2
     #   pre-commit
 requests==2.32.3
     # via sphinx
+scipy==1.14.1
+    # via
+    #   jax
+    #   jaxlib
 snowballstemmer==2.2.0
     # via sphinx
 soupsieve==2.6
     # via beautifulsoup4
 sphinx==6.2.1
     # via
-    #   -r requirements.in
+    #   -r support/requirements.in
     #   breathe
     #   furo
     #   myst-parser
@@ -107,7 +126,7 @@ sphinxcontrib-qthelp==2.0.0
 sphinxcontrib-serializinghtml==2.0.0
     # via sphinx
 sympy==1.13.3
-    # via -r requirements.in
+    # via -r support/testing-requirements.in
 tomli==2.0.2
     # via mypy
 typing-extensions==4.12.2
