+++ support/testing-requirements.in
@@ -1,0 +1,4 @@
+# Tests:
+numpy
+sympy
+jax
