+++ support/requirements.in
@@ -2,9 +2,6 @@
 cmake>=3.20
 ninja>=1.5
 mypy==1.9.0
-# Tests:
-numpy
-sympy
 # Documentation
 # We are stuck on an older sphinx until breathe resolves compatibility issues.
 # Such as: https://github.com/sphinx-doc/sphinx/issues/11490
