+++ examples/jax_photometric_alignment/README.md
@@ -1,0 +1,6 @@
+# jax_photometric_alignment
+
+To run this example:
+```bash
+python -m examples.jax_photometric_alignment.jax_photometric_alignment
+```
