+++ examples/CMakeLists.txt
@@ -4,6 +4,7 @@ add_subdirectory(bspline)
 add_subdirectory(c_generation)
 add_subdirectory(custom_types)
 add_subdirectory(imu_integration)
+add_subdirectory(jax_photometric_alignment)
 add_subdirectory(motion_planning)
 add_subdirectory(python_generation)
 add_subdirectory(quaternion_interpolation)
