+++ examples/python_generation/CMakeLists.txt
@@ -1,1 +1,2 @@
-add_python_test(python_generation.py)
+add_python_test(python_generation.py SOURCE_FILES
+                ${WF_EXAMPLES_DIR}/shared_expressions.py)
