+++ examples/jax_photometric_alignment/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_python_test(jax_photometric_alignment.py SOURCE_FILES utils.py
+                ${WF_EXAMPLES_DIR}/shared_expressions.py)
