+++ pyproject.toml
@@ -75,4 +75,4 @@ test-requires = ["numpy", "sympy", "jax"]
 [tool.isort]
 multi_line_output = 3
 include_trailing_comma = true
-known_first_party = ["wrenfold", "pywrenfold", "test_base"]
+known_first_party = ["wrenfold", "pywrenfold"]
