+++ components/core/wf/code_generation/function_description.h
@@ -46,11 +46,17 @@ class argument {
     return impl_->direction == argument_direction::optional_output;
   }
 
+  // Is this an input argument?
+  bool is_input() const noexcept { return impl_->direction == argument_direction::input; }
+
   // Argument direction.
   argument_direction direction() const noexcept { return impl_->direction; }
 
   // Position of this argument in the argument list.
   std::size_t index() const noexcept { return impl_->index; }
+
+  // Create symbolic input expressions that go with this argument.
+  std::variant<scalar_expr, matrix_expr, compound_expr> create_symbolic_input() const;
 
  private:
   // We share arguments in multiple places, and return them into python.
