+++ components/core/wf/code_generation/rust_code_generator.cc
@@ -3,7 +3,6 @@
 // For license information refer to accompanying LICENSE file.
 #include "wf/code_generation/rust_code_generator.h"
 
-#include "wf/code_generation/ast_formatters.h"
 #include "wf/code_generation/ast_visitor.h"
 #include "wf/utility/index_range.h"
 #include "wf/utility/overloaded_visit.h"
