+++ components/core/wf/external_function.cc
@@ -3,7 +3,6 @@
 // For license information refer to accompanying LICENSE file.
 #include "wf/external_function.h"
 
-#include "wf/code_generation/ast_formatters.h"
 #include "wf/expression_visitor.h"
 #include "wf/utility/overloaded_visit.h"
 
