+++ components/core/wf/code_generation/python_code_generator.cc
@@ -1,0 +1,460 @@
+// wrenfold symbolic code generator.
+// Copyright (c) 2024 Gareth Cross
+// For license information refer to accompanying LICENSE file.
+#include "wf/code_generation/python_code_generator.h"
+
+#include <ranges>
+#include <span>
+
+#include "wf/code_generation/ast_formatters.h"
+#include "wf/code_generation/ast_visitor.h"
+
+namespace wf {
+
+constexpr std::string_view python_string_from_float_width(
+    python_generator_float_width width) noexcept {
+  switch (width) {
+    case python_generator_float_width::float32: {
+      return "float32";
+    }
+    case python_generator_float_width::float64: {
+      return "float64";
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+constexpr static std::string_view python_string_from_cast_destination_type(
+    const numeric_primitive_type destination_type,
+    const python_generator_float_width float_width) noexcept {
+  switch (destination_type) {
+    case numeric_primitive_type::boolean:
+      return "bool";
+    case numeric_primitive_type::integral:
+      // Int32 because JAX does not support 64 by default. TODO: add a parameter for this.
+      return "int32";
+    case numeric_primitive_type::floating_point: {
+      return python_string_from_float_width(float_width);
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+constexpr static std::string_view python_module_prefix_from_target(
+    const python_generator_target target) noexcept {
+  switch (target) {
+    case python_generator_target::numpy: {
+      return "np";
+    }
+    case python_generator_target::pytorch: {
+      return "th";
+    }
+    case python_generator_target::jax: {
+      return "jnp";
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+inline std::string format_asarray(const python_generator_target target, const std::string_view arg,
+                                  const numeric_primitive_type expected_type,
+                                  const python_generator_float_width float_width,
+                                  const index_t rows, const index_t cols) {
+  return fmt::format(
+      "{}.asarray({}, dtype={}.{}).reshape([{}, {}])", python_module_prefix_from_target(target),
+      arg, python_module_prefix_from_target(target),
+      python_string_from_cast_destination_type(expected_type, float_width), rows, cols);
+}
+
+python_code_generator::python_code_generator(python_generator_target target,
+                                             python_generator_float_width float_width, int indent)
+    : target_(target), float_width_(float_width), indent_(static_cast<std::size_t>(indent)) {
+  if (indent < 1) {
+    throw wf::invalid_argument_error("Indentation must be >= 1. Provided value: {}", indent);
+  }
+}
+
+constexpr std::string_view python_matrix_type_from_target(python_generator_target target) noexcept {
+  switch (target) {
+    case python_generator_target::numpy: {
+      return "np.ndarray";
+    }
+    case python_generator_target::pytorch: {
+      return "th.Tensor";
+    }
+    case python_generator_target::jax: {
+      return "jnp.ndarray";
+    }
+  }
+  return "<INVALID ENUM VALUE>";
+}
+
+std::string python_code_generator::operator()(const scalar_type& scalar) const {
+  // Arguments can be passed as scalars, or as single-element arrays/tensors.
+  // So we annotate with either type.
+  // Once we drop Python 3.9 support, this can be a union written like: `X | Y`
+  // See: https://peps.python.org/pep-0604/
+  const auto matrix_alternative = python_matrix_type_from_target(target_);
+  if (target_ == python_generator_target::numpy) {
+    // Numpy treats float32, float64, bool, int64 as specific types.
[…]
