### 2024-11-15 df42515
Add example of generating KB camera model w/ JAX
