### 2024-11-15 #7303 @abougouffa
Hello @gareth-cross,

> Hi @abougouffa thank you for your review.

You're welcome,

> It looks like the checklist is complete.

Yep, I've completed the checklist. Sorry, I forgot to add a final comment. Anyway, I didn't find anything extra to mention. The paper is clear and follows the journal guidelines. And the repo is well organized, easy to install and use. I followed some examples provided in the repo, and I was able to reproduce the results.

I would like to felicitate you for this work. I will certainly use your library if I need to prototype or implement an algorithm with complex math expressions.

Thanks again!

A.B.
