### 2024-11-15 #7303 @boisgera
> Hi @abougouffa thank you for your review. It looks like the checklist is complete. Are there any other steps I should take @boisgera ?

Let me ping the 2nd reviewer (@ushu). I know he has been working on the review of your project, but AFAICT he has not yet created his checklist. Sorry for not being more reactive, I should have seen that @abougouffa had completed his checklist a while ago (unfortunately I don't get automatically notified when a reviewer checks a box).
