### 2024-11-15 7c79708
Add example of generating KB camera model w/ JAX (#294)
