+++ examples/jax_camera_model/jax_camera_model.py
@@ -1,0 +1,171 @@
+"""
+Code generate a camera model in JAX and test it numerically.
+"""
+
+import typing as T
+from types import SimpleNamespace
+
+import jax
+import jax.numpy as jnp
+import numpy as np
+
+from wrenfold import code_generation
+
+from ..shared_expressions import (
+    kb_camera_projection_with_jacobians,
+    kb_camera_unprojection_with_jacobians,
+)
+
+
+def get_test_camera_coeffs() -> T.List[T.Tuple[jnp.ndarray, float]]:
+    """Get test coefficients for some representative camera models."""
+    coeffs = [
+        ([0.08327424, 0.00852979, 0.00063325, 0.00017048], 1.99),
+        ([0.0, 0.0, 0.0, 0.0], jnp.pi / 2),
+        ([-4.16666666e-02, 5.20833056e-04, -3.09988000e-06, 1.06158708e-08], 1.41),
+        ([-1.66666587e-01, 8.33305775e-03, -1.98095183e-04, 2.60641004e-06], 0.99),
+    ]
+    return [(jnp.asarray(k).reshape([-1, 1]), max_radius) for (k, max_radius) in coeffs]
+
+
+def generate_kb_camera_model_functions() -> T.Tuple[T.Callable, T.Callable, T.Callable, T.Callable]:
+    """
+    Generate forward and backward projection functions for the Kannala-Brandt camera model.
+
+    We call ``generate_python``, which returns a callable that operates on JAX types.
+    """
+    project, _ = code_generation.generate_python(
+        func=kb_camera_projection_with_jacobians,
+        target=code_generation.PythonGeneratorTarget.JAX,
+    )
+    unproject, _ = code_generation.generate_python(
+        func=kb_camera_unprojection_with_jacobians,
+        target=code_generation.PythonGeneratorTarget.JAX,
+    )
+
+    # Batch over points (first argument), and then JIT with xla.
+    # We need to specify which derivatives we want here, since `vmap` does not know what to do with
+    # these boolean arguments.
+    project_batched = jax.vmap(
+        lambda *args: project(
+            *args,
+            compute_p_pixels_D_p_cam=True,
+            compute_p_pixels_D_K=True,
+            compute_p_pixels_D_coeffs=True,
+        ),
+        in_axes=(0, None, None),
+        out_axes=0,
+    )
+
+    unproject_batched = jax.vmap(
+        lambda *args: unproject(
+            *args,
+            compute_p_cam_D_p_pixels=True,
+            compute_p_cam_D_K=True,
+            compute_p_cam_D_coeffs=True,
+        ),
+        in_axes=(0, None, None),
+        out_axes=0,
+    )
+
+    project_no_jacobians = jax.vmap(
+        lambda *args: project(
+            *args,
+            compute_p_pixels_D_p_cam=False,
+            compute_p_pixels_D_K=False,
+            compute_p_pixels_D_coeffs=False,
+        )["p_pixels"],
+        in_axes=(0, None, None),
+        out_axes=0,
+    )
+
+    unproject_no_jacobians = jax.vmap(
+        lambda *args: unproject(
+            *args,
+            compute_p_cam_D_p_pixels=False,
+            compute_p_cam_D_K=False,
+            compute_p_cam_D_coeffs=False,
+        )["p_cam"],
+        in_axes=(0, None, None),
+        out_axes=0,
+    )
+
+    return (
+        jax.jit(project_batched),
+        jax.jit(unproject_batched),
+        jax.jit(project_no_jacobians),
[…]
