+++ examples/jax_camera_model/README.md
@@ -1,0 +1,6 @@
+# jax_camera_model
+
+To run this example:
+```bash
+python -m examples.jax_camera_model.jax_camera_model
+```
