+++ examples/jax_camera_model/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_python_test(jax_camera_model.py SOURCE_FILES
+                ${WF_EXAMPLES_DIR}/shared_expressions.py)
