### 2024-11-16 491b5d0
Add basic contributing.md (#295)
