### 2024-11-17 7536426
Add code coverage github action (#296)
