### 2024-11-17 86e1005
Add code_coverage.py script
