### 2024-11-17 2e69d1e
Add a few missing test cases on the core library
