+++ components/core/wf/expressions/matrix.cc
@@ -3,7 +3,7 @@
 // For license information refer to accompanying LICENSE file.
 #include "wf/expressions/matrix.h"
 
-#include "wf/expression_visitor.h"
+#include "wf/constants.h"
 #include "wf/expressions/addition.h"
 #include "wf/utility/assertions.h"
 
@@ -32,22 +32,6 @@ matrix matrix::get_block(const index_t row, const index_t col, const index_t nro
     data.push_back(get_unchecked(i + row, j + col));
   });
   return matrix(nrows, ncols, std::move(data));
-}
-
-void matrix::set_block(const index_t row, const index_t col, const index_t nrows,
-                       const index_t ncols, const matrix& block) {
-  if (row < 0 || row + nrows > rows_ || col < 0 || col + ncols > cols_) {
-    throw dimension_error(
-        "Block [position: ({}, {}), size: ({}, {})] is out of bounds for matrix of shape ({}, {})",
-        row, col, nrows, ncols, rows_, cols_);
-  }
-  if (block.rows() != nrows || block.cols() != ncols) {
-    throw dimension_error("Block shape ({}, {}) does not match requested shape ({}, {})",
-                          block.rows(), block.cols(), nrows, ncols);
-  }
-  iter_matrix(nrows, ncols, [&](index_t i, index_t j) {
-    set_unchecked(i + row, j + col, block.get_unchecked(i, j));
-  });
 }
 
 matrix matrix::transposed() const {
