### 2024-11-17 e2407bd
Add tree_formatter_test
