### 2024-11-17 54bd8eb
Add some new test coverage, add tests for tree_formatter (#297)
* Add code_coverage.py script

* Missing tests on the wrapper

* Add a few missing test cases on the core library

* Add tree_formatter_test
