+++ components/core/wf/expressions/addition.cc
@@ -7,7 +7,6 @@
 #include "wf/expressions/multiplication.h"
 #include "wf/expressions/numeric_expressions.h"
 #include "wf/expressions/special_constants.h"
-#include "wf/utility/overloaded_visit.h"
 
 namespace wf {
 
