+++ components/wrapper/pywrenfold/matrix_wrapper.cc
@@ -350,6 +350,11 @@ void wrap_matrix_operations(py::module_& m) {
           [](const matrix_expr& self,
              const std::variant<std::vector<scalar_expr>, matrix_expr>& vars,
              const bool use_abstract) {
+            if (self.rows() != 1 && self.cols() != 1) {
+              throw dimension_error(
+                  "Jacobian can only be computed on vectors. Received dimensions: [{}, {}]",
+                  self.rows(), self.cols());
+            }
             return jacobian(self.as_matrix().children(), span_from_variant(vars),
                             use_abstract ? non_differentiable_behavior::abstract
                                          : non_differentiable_behavior::constant);
