### 2024-11-21 d201986
Add cart double-pole dynamics + unit test
