+++ examples/cart_pole/cart_pole_dynamics.py
@@ -1,0 +1,141 @@
+"""
+Specify the dynamics of cart-pole system with a double-pendulum attached.
+"""
+
+import dataclasses
+import typing as T
+
+from wrenfold import code_generation, sym, type_annotations
+
+from .sympy_helpers import get_euler_lagrange_coefficients, get_mat_inverse
+
+
+@dataclasses.dataclass
+class CartPoleParamsSymbolic:
+    """Parameters of the cart-double-pole system."""
+
+    m_b: type_annotations.FloatScalar
+    m_1: type_annotations.FloatScalar
+    m_2: type_annotations.FloatScalar
+    l_1: type_annotations.FloatScalar
+    l_2: type_annotations.FloatScalar
+    g: type_annotations.FloatScalar
+
+
+def get_cart_double_pole_dynamics() -> T.Callable:
+    """
+    Return a symbolic function that evaluates the dynamics of a cart-mounted double pendulum.
+
+    We derive the necessary expressions symbolically, and then create the function
+    `cart_double_pole_dynamics` that specifies exactly what we want code-generated.
+    """
+    t = sym.symbols("t", real=True)
+
+    # Position of the base + angles as a function of time.
+    # We create symbolic functions to represent these, and then substitute in the function arguments
+    # below.
+    b_x = sym.Function("b_x")(t)
+    th_1 = sym.Function("th_1")(t)
+    th_2 = sym.Function("th_2")(t)
+
+    # First derivatives of the base + angles:
+    b_x_dot = b_x.diff(t)
+    th_1_dot = th_1.diff(t)
+    th_2_dot = th_2.diff(t)
+
+    # Mass of the base, and two weights + lever arm lengths:
+    m_b, m_1, m_2, l_1, l_2 = sym.symbols("m_b, m_1, m_2, l_1, l_2", real=True, positive=True)
+
+    # Gravity:
+    g = sym.symbols("g", real=True)
+
+    # Positions of base, and two weights.
+    b = sym.vector(b_x, 0)
+    p_1 = b + sym.vector(sym.cos(th_1), sym.sin(th_1)) * l_1
+    p_2 = p_1 + sym.vector(sym.cos(th_2), sym.sin(th_2)) * l_2
+
+    b_dot = b.diff(t)
+    p_1_dot = p_1.diff(t)
+    p_2_dot = p_2.diff(t)
+
+    # Compute kinetic energy. This is the sum of (1/2)*m*v^2 for all pieces.
+    half = 1 / sym.integer(2)
+    T: sym.Expr = (
+        half * m_b * b_dot.squared_norm() + half * m_1 * p_1_dot.squared_norm() +
+        half * m_2 * p_2_dot.squared_norm())
+
+    # Simplify this a bit by eliminating: cos^2(x) + sin^2(x) --> 1
+    T = (
+        T.distribute().collect([m_1, m_2, l_1, l_2, th_1_dot, th_2_dot]).subs(
+            (sym.cos(th_1) ** 2 + sym.sin(th_1) ** 2) / 2, half).subs(
+                (sym.cos(th_2) ** 2 + sym.sin(th_2) ** 2) / 2, half))
+
+    # Compute potential energy. This is the sum of m*g*y for all pieces.
+    V = g * m_1 * p_1[1] + g * m_2 * p_2[1]
+
+    # The lagrangian:
+    L: sym.Expr = T - V
+
+    # Canonical momenta:
+    q_b = L.diff(b_x_dot)
+    q_th_1 = L.diff(th_1_dot)
+    q_th_2 = L.diff(th_2_dot)
+
+    # Form the Euler-Lagrange equations (each of these is equal to zero).
[…]
