+++ examples/cart_pole/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_python_test(cart_pole_test.py SOURCE_FILES cart_pole_dynamics.py
+                sympy_helpers.py)
