+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "wrenfold"
-version = "0.1.1"
+version = "0.2.1"
 requires-python = ">=3.9"
 authors = [
     { name = "Gareth Cross", email = "gcross.code@icloud.com" },
