+++ support/conda/wrenfold/meta.yaml
@@ -1,4 +1,4 @@
-{% set version = "0.1.1" %}
+{% set version = "0.2.1" %}
 
 package:
   name: wrenfold
