+++ examples/cart_pole/cart_pole_demo/src/main.rs
@@ -1,0 +1,290 @@
+//! wrenfold symbolic code generator.
+//! Copyright (c) 2024 Gareth Cross
+//! For license information refer to accompanying LICENSE file.
+use egui::{pos2, vec2, Color32, Rect, Shape, Stroke, Ui};
+use nalgebra as na;
+use rand::prelude::Distribution;
+
+mod generated;
+
+#[derive(Debug, Clone, Copy)]
+pub struct CartPoleParams {
+    pub m_b: f64,
+    pub m_1: f64,
+    pub m_2: f64,
+    pub l_1: f64,
+    pub l_2: f64,
+    pub g: f64,
+}
+
+impl Default for CartPoleParams {
+    fn default() -> Self {
+        Self {
+            m_b: 1.0,
+            m_1: 0.25,
+            m_2: 0.20,
+            l_1: 0.30,
+            l_2: 0.15,
+            g: 9.81,
+        }
+    }
+}
+
+fn runge_kutta_4th_order<const DIM: usize, F>(
+    x: &na::SMatrix<f64, DIM, 1>,
+    h: f64,
+    f: F,
+) -> na::SMatrix<f64, DIM, 1>
+where
+    F: Fn(&na::SMatrix<f64, DIM, 1>) -> na::SMatrix<f64, DIM, 1>,
+{
+    let k1 = f(x);
+    let k2 = f(&(x + k1 * h / 2.0));
+    let k3 = f(&(x + k2 * h / 2.0));
+    let k4 = f(&(x + k3 * h));
+    x + (h / 6.0) * (k1 + k2 * 2.0 + k3 * 2.0 + k4)
+}
+
+/// Main struct of the app.
+pub struct App {
+    params: CartPoleParams,
+    state: na::Vector6<f64>,
+    sim_rate: f64,
+    previous_update_time: Option<std::time::Instant>,
+}
+
+impl App {
+    pub fn new(cc: &eframe::CreationContext<'_>) -> Self {
+        egui_extras::install_image_loaders(&cc.egui_ctx);
+        Self {
+            params: Default::default(),
+            state: na::Vector6::zeros(),
+            sim_rate: 1.0,
+            previous_update_time: None,
+        }
+    }
+
+    /// Step the simulator forward.
+    fn step_sim(&mut self) {
+        let current_time = std::time::Instant::now();
+        if let Some(previous_update_time) = self.previous_update_time {
+            let mut dt = current_time
+                .duration_since(previous_update_time)
+                .as_secs_f64()
+                * self.sim_rate;
+
+            // Break interval into sub-intervals to improve integration accuracy.
+            const INTERNAL_DT: f64 = 0.0001;
+            while dt > 0.0 {
+                self.sub_step_sim(dt.min(INTERNAL_DT));
+                dt -= INTERNAL_DT;
+            }
+        }
+        self.previous_update_time = Some(current_time);
+    }
+
+    fn sub_step_sim(&mut self, dt: f64) {
+        self.state = runge_kutta_4th_order(&self.state, dt, |x| {
+            let mut x_dot = na::Vector6::zeros();
+            generated::cart_double_pole_dynamics(
+                &self.params,
+                x,
+                &mut x_dot,
+                Option::<&mut f64>::None,
+                Option::<&mut na::Matrix6<f64>>::None,
+            );
+            x_dot
+        });
+    }
+
+    /// Randomize the initial conditions.
+    fn reset_state(&mut self) {
+        let dist = rand::distributions::Uniform::new(-std::f64::consts::PI, std::f64::consts::PI);
+        let mut rng = rand::thread_rng();
+        self.state[0] = 0.0; //  initial position
+        self.state[1] = dist.sample(&mut rng);
[…]
