### 2024-11-23 4152c41
Change PyProject build type to release
