+++ CMakeLists.txt
@@ -28,7 +28,7 @@ option(WF_BUILD_EXTRAS "Build tests, benchmarks, examples, and docs."
        ${WF_BUILD_EXTRAS_DEFAULT})
 option(WF_STUBS_REQUIRED "Fail if stubgen cannot be found." OFF)
 option(WF_DOCS_REQUIRED "Fail if sphinx/doxygen cannot be found." OFF)
-option(WF_CODE_COVERAGE "Build in code coverag mode." OFF)
+option(WF_CODE_COVERAGE "Build in code coverage mode." OFF)
 
 # Add third party code
 add_subdirectory(dependencies)
@@ -66,6 +66,9 @@ if(WF_CODE_COVERAGE AND CMAKE_CXX_COMPILER_ID MATCHES "GNU|Clang")
   # Turn off optimization, add debug info, and enable coverage.
   target_compile_options(wf_coverage_config INTERFACE -O0 -g --coverage)
   target_link_options(wf_coverage_config INTERFACE --coverage)
+  if(NOT CMAKE_BUILD_TYPE STREQUAL "Debug")
+    message(FATAL_ERROR "Build type must be Debug to build code coverage.")
+  endif()
 endif()
 
 # Custom target that groups together all the rust code-generation.
