### 2024-11-23 a2bf2b6
Change PyProject build type to release (#299)
