+++ pyproject.toml
@@ -46,15 +46,11 @@ build-backend = "scikit_build_core.build"
 cmake.version = ">=3.20"
 # It seems like setting `build-type` is insufficient on Windows, and we fall back to Debug.
 # Place it in the `args` list to force a release build.
-cmake.args = ["-G", "Ninja", "-Wno-deprecated", "-DCMAKE_BUILD_TYPE=RelWithDebInfo", "-DWF_STUBS_REQUIRED=ON"]
-cmake.verbose = true
-cmake.build-type = "RelWithDebInfo"
-
+cmake.args = ["-G", "Ninja", "-Wno-deprecated", "-DCMAKE_BUILD_TYPE=Release", "-DWF_STUBS_REQUIRED=ON"]
+cmake.build-type = "Release"
 ninja.version = ">=1.5"
 ninja.make-fallback = false
-
 logging.level = "WARNING"
-
 wheel.packages = ["components/python/wrenfold"]
 wheel.license-files = [
     "LICENSE",
@@ -62,6 +58,7 @@ wheel.license-files = [
     "dependencies/fmt/LICENSE",
     "dependencies/pybind11/LICENSE"
 ]
+build.verbose = true
 
 [tool.cibuildwheel]
 skip = ["*musllinux_*", "*-win32", "*_i686", "pp*"]
