### 2024-11-23 e1d2486
Update version to v0.2.1 (#298)
* Update version to v0.2.1

* Fix compilation error on clang
