### 2024-11-23 7dd2862
Add dissipative forces to example
