+++ examples/cart_pole/cart_pole_demo/src/generated.rs
@@ -10,13 +10,14 @@ where
   T4: wrenfold_traits::OutputSpan2D<6, 6, ValueType = f64>,
 {
   // Operation counts:
-  // add: 87
-  // branch: 2
-  // call: 4
-  // divide: 2
-  // multiply: 283
-  // negate: 16
-  // total: 394
+  // add: 158
+  // branch: 11
+  // call: 7
+  // compare: 5
+  // divide: 5
+  // multiply: 444
+  // negate: 23
+  // total: 653
   
   let v00017: f64 = x.get(2, 0);
   let v00004: f64 = x.get(1, 0);
@@ -28,102 +29,292 @@ where
   let v00002: f64 = x.get(5, 0);
   let v00008: f64 = params.l_1;
   let v00001: f64 = x.get(4, 0);
-  let v00901: f64 = v00002 * v00014;
-  let v00096: f64 = v00016 * v00018 + v00005 * v00020;
-  let v01135: f64 = -v00005;
+  let v01401: f64 = v00002 * v00014;
+  let v00175: f64 = v00016 * v00018 + v00005 * v00020;
+  let v01773: f64 = -v00005;
   let v00000: f64 = x.get(3, 0);
-  let v00895: f64 = v00001 * v00008;
+  let v01777: f64 = -v00020;
+  let v01396: f64 = v00001 * v00008;
   let v00026: f64 = params.g;
   let v00011: f64 = params.m_2;
   let v00007: f64 = params.m_1;
-  let v00038: f64 = params.m_b;
+  let v00080: f64 = params.m_b;
   if let Some(energy) = energy {
-    let v00084: f64 = 0.5f64;
-    let v00323: f64 = v00000 * v01135;
-    let v00324: f64 = v00084 * v00895;
-    let v00919: f64 = v00005 * v00008;
-    let v01088: f64 = v00084 * (v00000 * v00000);
-    *energy = v00038 * v01088 + v00007 * (((v00323 + v00324) * v00895 + v01088) + v00026 * v00919) + v00011 * (((v00324 + (v00323 + v00096 * v00901)) * v00895 + (v00000 * -v00020 + v00084 * v00901) * v00901 + v01088) + v00026 * (v00919 + v00014 * v00020));
+    let v00163: f64 = 0.5f64;
+    let v00617: f64 = v00000 * v01773;
+    let v00618: f64 = v00163 * v01396;
+    let v01418: f64 = v00005 * v00008;
+    let v01704: f64 = v00163 * (v00000 * v00000);
+    *energy = v00080 * v01704 + v00007 * (((v00617 + v00618) * v01396 + v01704) + v00026 * v01418) + v00011 * (((v00618 + (v00617 + v00175 * v01401)) * v01396 + (v00000 * v01777 + v00163 * v01401) * v01401 + v01704) + v00026 * (v01418 + v00014 * v00020));
   }
-  let v01134: f64 = -v00011;
-  let v01131: f64 = v00007 + v00011;
-  let v00039: f64 = v00038 + v01131;
-  let v01143: f64 = -v01131;
-  let v00533: f64 = v01131 * v01135;
-  let v00914: f64 = v00002 * v00002;
-  let v00887: f64 = v00011 * v00014;
-  let v01000: f64 = v00016 * v00020;
-  let v00564: f64 = v00039 + v00020 * v00020 * v01134;
-  let v01080: f64 = v01135 * v01143;
-  let v00958: f64 = v00039 * v00096;
-  let v00907: f64 = v00001 * v00001;
-  let v00892: f64 = v00008 * v01131;
-  let v00400: f64 = v00020 * -v00016 + v00005 * v00018;
[…]
