+++ examples/cart_pole/cart_pole_dynamics.py
@@ -15,12 +15,17 @@ from .sympy_helpers import get_euler_lagrange_coefficients, get_mat_inverse
 class CartPoleParamsSymbolic:
     """Parameters of the cart-double-pole system."""
 
-    m_b: type_annotations.FloatScalar
-    m_1: type_annotations.FloatScalar
-    m_2: type_annotations.FloatScalar
-    l_1: type_annotations.FloatScalar
-    l_2: type_annotations.FloatScalar
-    g: type_annotations.FloatScalar
+    m_b: type_annotations.FloatScalar  # Mass of the base/cart (kg)
+    m_1: type_annotations.FloatScalar  # Mass of the first pole (kg).
+    m_2: type_annotations.FloatScalar  # Mass of the second pole (kg).
+    l_1: type_annotations.FloatScalar  # Length of the first pole (m).
+    l_2: type_annotations.FloatScalar  # Length of the second pole (m).
+    g: type_annotations.FloatScalar  # Gravity (m/s^2).
+    mu_b: type_annotations.FloatScalar  # Coefficient of static friction.
+    v_mu_b: type_annotations.FloatScalar  # Velocity cutoff of smooth friction model (m/s).
+    c_d: type_annotations.FloatScalar  # Aero-coefficient of poles (N / (m/s)^2).
+    x_s: type_annotations.FloatScalar  # Bounding spring position (m).
+    k_s: type_annotations.FloatScalar  # Bounding spring constant (N / m).
 
 
 def get_cart_double_pole_dynamics() -> T.Callable:
@@ -49,6 +54,16 @@ def get_cart_double_pole_dynamics() -> T.Callable:
 
     # Gravity:
     g = sym.symbols("g", real=True)
+
+    # Friction coefficient on the base, and the cutoff velocity of the smooth Coulomb model.
+    mu_b, v_mu_b = sym.symbols("mu_b, v_mu_b", real=True)
+
+    # Air drag coefficient on the mass:
+    # This is summarized as: rho * C_d * A, where rho is air density, and A is cross-sectional area.
+    c_d = sym.symbols("c_d_1", real=True)
+
+    # Position and spring coefficient on the boundary of the workspace.
+    x_s, k_s = sym.symbols("x_s, k_s", real=True)
 
     # Positions of base, and two weights.
     b = sym.vector(b_x, 0)
@@ -82,10 +97,26 @@ def get_cart_double_pole_dynamics() -> T.Callable:
     q_th_1 = L.diff(th_1_dot)
     q_th_2 = L.diff(th_2_dot)
 
+    # Dissipative force due to friction on the base.
+    F_friction_base = (-mu_b * (m_1 + m_2 + m_b) * g * sym.tanh(b_x_dot / sym.max(v_mu_b, 1.0e-6)))
+
+    # Dissipative _power_ due to air drag on the pendulum mass.
+    # We use a `where` statement to guard against a singularity in the Jacobian.
+    D_air_mass_1 = ((sym.integer(1) / 6) * c_d * sym.where(p_1_dot.squared_norm() > 0,
+                                                           p_1_dot.norm() ** 3, 0))
+    D_air_mass_2 = ((sym.integer(1) / 6) * c_d * sym.where(p_2_dot.squared_norm() > 0,
+                                                           p_2_dot.norm() ** 3, 0))
+    D_air_mass = D_air_mass_1 + D_air_mass_2
+
+    # External force from the boundary spring:
+    F_s_right = -k_s * sym.max(0, b_x - x_s)
+    F_s_left = k_s * sym.max(0, -x_s - b_x)
+
[…]
