### 2024-11-24 cbfb849
Add arrow-key functionality to move the cart
