+++ examples/cart_pole/cart_pole_demo/src/generated.rs
@@ -3,318 +3,319 @@
 
 #[inline]
 #[allow(non_snake_case, clippy::unused_unit, clippy::collapsible_else_if, clippy::needless_late_init, unused_variables, unused_parens)]
-pub fn cart_double_pole_dynamics<T1, T2, T4, >(params: &crate::CartPoleParams, x: &T1, x_dot: &mut T2, energy: Option<&mut f64>, J_x: Option<&mut T4>) -> ()
+pub fn cart_double_pole_dynamics<T1, T3, T5, >(params: &crate::CartPoleParams, x: &T1, u: f64, x_dot: &mut T3, energy: Option<&mut f64>, J_x: Option<&mut T5>) -> ()
 where
   T1: wrenfold_traits::Span2D<6, 1, ValueType = f64>,
-  T2: wrenfold_traits::OutputSpan2D<6, 1, ValueType = f64>,
-  T4: wrenfold_traits::OutputSpan2D<6, 6, ValueType = f64>,
+  T3: wrenfold_traits::OutputSpan2D<6, 1, ValueType = f64>,
+  T5: wrenfold_traits::OutputSpan2D<6, 6, ValueType = f64>,
 {
   // Operation counts:
-  // add: 158
+  // add: 159
   // branch: 11
   // call: 7
   // compare: 5
   // divide: 5
   // multiply: 444
   // negate: 23
-  // total: 653
+  // total: 654
   
-  let v00017: f64 = x.get(2, 0);
-  let v00004: f64 = x.get(1, 0);
-  let v00020: f64 = (v00017).sin();
-  let v00005: f64 = (v00004).sin();
-  let v00018: f64 = (v00017).cos();
-  let v00016: f64 = (v00004).cos();
-  let v00014: f64 = params.l_2;
+  let v00055: f64 = x.get(2, 0);
+  let v00034: f64 = x.get(1, 0);
+  let v00056: f64 = (v00055).sin();
+  let v00038: f64 = (v00034).sin();
+  let v00062: f64 = (v00055).cos();
+  let v00035: f64 = (v00034).cos();
+  let v00057: f64 = params.l_2;
   let v00002: f64 = x.get(5, 0);
-  let v00008: f64 = params.l_1;
+  let v00036: f64 = params.l_1;
   let v00001: f64 = x.get(4, 0);
-  let v01401: f64 = v00002 * v00014;
-  let v00175: f64 = v00016 * v00018 + v00005 * v00020;
-  let v01773: f64 = -v00005;
+  let v01402: f64 = v00002 * v00057;
+  let v00176: f64 = v00035 * v00062 + v00038 * v00056;
+  let v01774: f64 = -v00038;
   let v00000: f64 = x.get(3, 0);
-  let v01777: f64 = -v00020;
-  let v01396: f64 = v00001 * v00008;
-  let v00026: f64 = params.g;
-  let v00011: f64 = params.m_2;
+  let v01778: f64 = -v00056;
+  let v01397: f64 = v00001 * v00036;
+  let v00019: f64 = params.g;
+  let v00008: f64 = params.m_2;
   let v00007: f64 = params.m_1;
-  let v00080: f64 = params.m_b;
+  let v00006: f64 = params.m_b;
   if let Some(energy) = energy {
-    let v00163: f64 = 0.5f64;
-    let v00617: f64 = v00000 * v01773;
-    let v00618: f64 = v00163 * v01396;
-    let v01418: f64 = v00005 * v00008;
-    let v01704: f64 = v00163 * (v00000 * v00000);
[…]
