### 2024-11-25 a5b6452
Add cart-pole/double-pendulum to the examples (#300)
* Add cart double-pole dynamics + unit test

* Add cart-pole example

* Add dissipative forces to example

* Add references to documentation

* Add arrow-key functionality to move the cart
