+++ examples/cart_pole/cart_pole_dynamics.py
@@ -1,0 +1,192 @@
+"""
+Specify the dynamics of cart-pole system with a double-pendulum attached.
+"""
+
+import dataclasses
+import sys
+import typing as T
+
+from wrenfold import code_generation, sym, type_annotations, type_info
+
+from .sympy_helpers import get_euler_lagrange_coefficients, get_mat_inverse
+
+
+@dataclasses.dataclass
+class CartPoleParamsSymbolic:
+    """Parameters of the cart-double-pole system."""
+
+    m_b: type_annotations.FloatScalar  # Mass of the base/cart (kg)
+    m_1: type_annotations.FloatScalar  # Mass of the first pole (kg).
+    m_2: type_annotations.FloatScalar  # Mass of the second pole (kg).
+    l_1: type_annotations.FloatScalar  # Length of the first pole (m).
+    l_2: type_annotations.FloatScalar  # Length of the second pole (m).
+    g: type_annotations.FloatScalar  # Gravity (m/s^2).
+    mu_b: type_annotations.FloatScalar  # Coefficient of static friction.
+    v_mu_b: type_annotations.FloatScalar  # Velocity cutoff of smooth friction model (m/s).
+    c_d: type_annotations.FloatScalar  # Aero-coefficient of poles (N / (m/s)^2).
+    x_s: type_annotations.FloatScalar  # Bounding spring position (m).
+    k_s: type_annotations.FloatScalar  # Bounding spring constant (N / m).
+
+
+def get_cart_double_pole_dynamics() -> T.Callable:
+    """
+    Return a symbolic function that evaluates the dynamics of a cart-mounted double pendulum.
+
+    We derive the necessary expressions symbolically, and then create the function
+    `cart_double_pole_dynamics` that specifies exactly what we want code-generated.
+    """
+    t = sym.symbols("t", real=True)
+
+    # Position of the base + angles as a function of time.
+    # We create symbolic functions to represent these, and then substitute in the function arguments
+    # below.
+    b_x = sym.Function("b_x")(t)
+    th_1 = sym.Function("th_1")(t)
+    th_2 = sym.Function("th_2")(t)
+
+    # First derivatives of the base + angles:
+    b_x_dot = b_x.diff(t)
+    th_1_dot = th_1.diff(t)
+    th_2_dot = th_2.diff(t)
+
+    # Mass of the base, and two weights + lever arm lengths:
+    m_b, m_1, m_2, l_1, l_2 = sym.symbols("m_b, m_1, m_2, l_1, l_2", real=True, positive=True)
+
+    # Gravity:
+    g = sym.symbols("g", real=True)
+
+    # Control input on the base:
+    u_b = sym.symbols("u_b", real=True)
+
+    # Friction coefficient on the base, and the cutoff velocity of the smooth Coulomb model.
+    mu_b, v_mu_b = sym.symbols("mu_b, v_mu_b", real=True)
+
+    # Air drag coefficient on the mass:
+    # This is summarized as: rho * C_d * A, where rho is air density, and A is cross-sectional area.
+    c_d = sym.symbols("c_d_1", real=True)
+
+    # Position and spring coefficient on the boundary of the workspace.
+    x_s, k_s = sym.symbols("x_s, k_s", real=True)
+
+    # Positions of base, and two weights.
+    b = sym.vector(b_x, 0)
+    p_1 = b + sym.vector(sym.cos(th_1), sym.sin(th_1)) * l_1
+    p_2 = p_1 + sym.vector(sym.cos(th_2), sym.sin(th_2)) * l_2
+
+    b_dot = b.diff(t)
+    p_1_dot = p_1.diff(t)
[…]
