### 2024-11-25 #7303 @boisgera
Hi!

> @boisgera am I supposed to put my remarks below in the PR ? Or in the [orignal repo](https://github.com/wrenfold/wrenfold) ? The whole process is pretty new to me an not obvious (yet) 😅

Most of the time, you create issues in the original repo and you mention [the JOSS review issue](https://github.com/openjournals/joss-reviews/issues/7303) so that we can track everything from here. The JOSS issue is the "meta-issue" if you wish.

And of course if there is something that does not fit into a project issue (e.g. a question about the JOSS process), you may add a comment here, for myself and/or the project author.
