### 2024-11-27 44f1aba
Add apply_preamble to the python code generator
