### 2024-11-28 6e85985
Add apply_preamble to the python code generator (#302)
