### 2024-11-29 07897a8
Do not set git hash at build time if we failed to obtain it
