+++ components/wrapper/CMakeLists.txt
@@ -1,27 +1,28 @@
-# Custom target for all stubs:
-add_custom_target(wf_stubs ALL)
+# Define a property for the list of submodules in our top level C++ module.
+set(PYTHON_SUBMODULES_DOC
+    "List of submodules built into the C++ wrapper module.")
+define_property(
+  TARGET
+  PROPERTY PYTHON_SUBMODULES
+  BRIEF_DOCS ${PYTHON_SUBMODULES_DOC}
+  FULL_DOCS ${PYTHON_SUBMODULES_DOC})
 
-# Find stubgen:
-if(NOT DEFINED STUBGEN_PATH OR STUBGEN_PATH STREQUAL "STUBGEN_PATH-NOTFOUND")
-  message(STATUS "Looking for stubgen...")
-  if(WF_STUBS_REQUIRED)
-    find_program(STUBGEN_PATH stubgen REQUIRED)
-  else()
-    find_program(STUBGEN_PATH stubgen)
-  endif()
-  if(DEFINED STUBGEN_PATH AND NOT STUBGEN_PATH STREQUAL "STUBGEN_PATH-NOTFOUND")
-    message(STATUS "stubgen path: ${STUBGEN_PATH}")
-  endif()
+if(WF_BUILD_STUBS)
+  add_custom_target(wf_stubs ALL)
 endif()
 
-# Function for defining the wrapper module.
+# Define a target for the C++ wrapper module.
 function(add_wrapper_module)
   set(oneValueArgs NAME)
-  set(multiValueArgs "SOURCES;SUB_MODULE_NAMES")
+  set(multiValueArgs "SOURCES;SUBMODULE_NAMES")
   cmake_parse_arguments(MODULE "" "${oneValueArgs}" "${multiValueArgs}" ${ARGN})
 
   # Add pybind11 module
   pybind11_add_module(${MODULE_NAME} ${MODULE_SOURCES})
+  set_property(
+    TARGET ${MODULE_NAME}
+    APPEND
+    PROPERTY PYTHON_SUBMODULES ${MODULE_SUBMODULE_NAMES})
 
   # Depend on the static library
   target_link_libraries(${MODULE_NAME} PRIVATE wf_core)
@@ -53,53 +54,12 @@ function(add_wrapper_module)
   endif()
 
   # For each submodule, define a compile-time constant for its name.
-  foreach(SUB_MODULE_NAME ${MODULE_SUB_MODULE_NAMES})
-    string(TOUPPER ${SUB_MODULE_NAME} SUB_MODULE_NAME_UPPER)
+  foreach(SUBMODULE_NAME ${MODULE_SUBMODULE_NAMES})
+    string(TOUPPER ${SUBMODULE_NAME} SUBMODULE_NAME_UPPER)
     target_compile_definitions(
       ${MODULE_NAME}
-      PRIVATE
-        "PY_SUBMODULE_NAME_${SUB_MODULE_NAME_UPPER}=\"${SUB_MODULE_NAME}\"")
+      PRIVATE "PY_SUBMODULE_NAME_${SUBMODULE_NAME_UPPER}=\"${SUBMODULE_NAME}\"")
   endforeach()
-
-  # Copy from Debug or Release into the current binary dir. We need to do this
-  # because msbuild puts libraries in sub-directories, which breaks our ability
-  # to import them from the cmake binary dir.
-  if(${CMAKE_GENERATOR} MATCHES "Visual Studio.+")
-    add_custom_command(
-      TARGET ${MODULE_NAME}
-      POST_BUILD
[…]
