+++ support/conda/wrenfold/meta.yaml
@@ -21,7 +21,6 @@ requirements:
     # Use cmake<3.28 to avoid clang-scan-deps dependency on OSX (introduced with CMake 3.28).
     - cmake >=3.20,<3.28
     - ninja >=1.5
-    - mypy ==1.9.0
     - pip
   run:
     - python
