### 2024-12-07 d56d62f
Stubs are generated into the repo and committed to components/wrapper/stubs (#305)
