+++ pyproject.toml
@@ -36,9 +36,7 @@ Issues = "https://github.com/wrenfold/wrenfold/issues"
 requires = [
     "scikit-build-core==0.10.7",
     "cmake>=3.20",
-    "ninja>=1.5",
-    # mypy required for stubgen.
-    "mypy==1.9.0",
+    "ninja>=1.5"
 ]
 build-backend = "scikit_build_core.build"
 
@@ -46,11 +44,11 @@ build-backend = "scikit_build_core.build"
 cmake.version = ">=3.20"
 # It seems like setting `build-type` is insufficient on Windows, and we fall back to Debug.
 # Place it in the `args` list to force a release build.
-cmake.args = ["-G", "Ninja", "-Wno-deprecated", "-DCMAKE_BUILD_TYPE=Release", "-DWF_STUBS_REQUIRED=ON"]
+cmake.args = ["-G", "Ninja", "-Wno-deprecated", "-DCMAKE_BUILD_TYPE=Release"]
 cmake.build-type = "Release"
 ninja.version = ">=1.5"
 ninja.make-fallback = false
-logging.level = "WARNING"
+logging.level = "INFO"
 wheel.packages = ["components/python/wrenfold"]
 wheel.license-files = [
     "LICENSE",
@@ -73,3 +71,4 @@ test-requires = ["numpy", "sympy", "jax"]
 multi_line_output = 3
 include_trailing_comma = true
 known_first_party = ["wrenfold", "pywrenfold", "test_base"]
+skip_glob = ["components/wrapper/stubs/*"]
