### 2024-12-07 dfe400b
Remove old python_generation example
