+++ examples/python_generation/python_generation.py
@@ -1,411 +1,0 @@
-"""
-An example of a custom code-generator implemented entirely in python.
-
-NOTE: This example is obsolete because python generation is now built-into the
-wrenfold framework. It will be removed in short order. See `c_generation` for
-an example of adding a new language.
-"""
-import dataclasses
-import typing as T
-import unittest
-
-import numpy as np
-
-from wrenfold import (
-    ast,
-    code_generation,
-    geometry,
-    sym,
-    type_annotations,
-    type_info,
-)
-from wrenfold.enumerations import (
-    RelationalOperation,
-    StdMathFunction,
-    SymbolicConstant,
-)
-
-
-@dataclasses.dataclass
-class SimParamsSymbolic:
-    """
-    A sample struct with some parameters of a simulation. This is the symbolic version.
-    We use this to test code-generation of custom types.
-    """
-    mass: type_annotations.FloatScalar
-    drag_coefficient: type_annotations.FloatScalar
-    gravity: type_annotations.FloatScalar
-
-
-@dataclasses.dataclass
-class SimParams:
-    """A sample struct with some parameters of a simulation."""
-    mass: float
-    drag_coefficient: float
-    gravity: float
-
-
-# noinspection PyMethodMayBeStatic
-class PythonCodeGenerator(code_generation.BaseGenerator):
-    """
-    A custom code-generator that emits python code. This could be used quickly try out a generated
-    function in an ipython notebook, for example. The code emitted by this generator is also
-    suitable for use with numba.
-
-    Caveat: Because python doesn't have an exact equivalent of "output arguments" from C++, we
-    convert output arguments into a tuple of return values. For optional outputs, we instead
-    generate a boolean argument that specifies whether to compute the optional value.
-    """
-
-    def __init__(self, indent: int = 2):
-        super().__init__()
-        assert indent > 0, f'indent = {indent}'
-        self._indent: str = ' ' * indent
-
-    def _indent_and_join(self, lines: T.Iterable[str]) -> str:
-        lines_split = []
-        for line in lines:
-            lines_split.extend(line.splitlines())
-        return self._indent + f'\n{self._indent}'.join(lines_split)
-
-    @staticmethod
-    def _format_numeric_type(t: type_info.NumericType) -> str:
-        if t == type_info.NumericType.Bool:
-            return 'bool'
-        elif t == type_info.NumericType.Integer:
-            return 'int'
-        elif t == type_info.NumericType.Float:
-            return 'float'
-        else:
-            raise NotImplementedError(f"Unsupported type: {t}")
-
-    def format_scalar_type(self, t: type_info.ScalarType) -> str:
-        return self._format_numeric_type(t.numeric_type)
-
-    def format_matrix_type(self, _: type_info.MatrixType) -> str:
-        return 'np.ndarray'
-
-    def format_custom_type(self, custom: type_info.CustomType) -> str:
-        if custom.python_type == SimParamsSymbolic:
-            return "SimParams"
-        return custom.name
-
-    def format_add(self, add: ast.Add) -> str:
-        return ' + '.join(self.format(x) for x in add.args)
-
-    def format_assign_output_matrix(self, mat: ast.AssignOutputMatrix) -> str:
-        mat_type = mat.arg.type
-        assert isinstance(mat_type, type_info.MatrixType)
-
-        args = ",\n  ".join([self.format(arg) for arg in mat.value.args])
-        return f'{mat.arg.name} = np.array([\n  {args}\n], dtype=np.float64).reshape({mat_type.rows}, {mat_type.cols})'
-
[…]
