+++ examples/python_generation/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_python_test(python_generation.py)
