### 2024-12-07 b2ec74d
Remove old python_generation example (#306)
