+++ examples/CMakeLists.txt
@@ -7,7 +7,6 @@ add_subdirectory(custom_types)
 add_subdirectory(imu_integration)
 add_subdirectory(jax_camera_model)
 add_subdirectory(motion_planning)
-add_subdirectory(python_generation)
 add_subdirectory(quaternion_interpolation)
 add_subdirectory(rosenbrock)
 add_subdirectory(rotation_error)
