### 2024-12-13 c13923b
Make NumPy the default python target
