### 2024-12-13 6d00c3c
Make NumPy the default python target (#308)
