### 2024-12-21 c43a42b
Update MyPy version in requirements
