+++ requirements.txt
@@ -61,7 +61,7 @@ ml-dtypes==0.5.0
     #   jaxlib
 mpmath==1.3.0
     # via sympy
-mypy==1.9.0
+mypy==1.14.0
     # via -r support/requirements.in
 mypy-extensions==1.0.0
     # via mypy
