### 2024-12-22 6260258
Update MyPy version in requirements (#309)
