+++ support/requirements.in
@@ -1,7 +1,7 @@
 # Building core library+wrapper:
 cmake>=3.20
 ninja>=1.5
-mypy==1.9.0
+mypy==1.14.0
 # Documentation
 # We are stuck on an older sphinx until breathe resolves compatibility issues.
 # Such as: https://github.com/sphinx-doc/sphinx/issues/11490
