### 2024-12-29 4458481
Add better documentation on derivatives and python code generation
