### 2024-12-29 0ae7c48
Add better documentation on derivatives and python code generation (#310)
