+++ components/wrapper/pywrenfold/compound_expression_wrapper.cc
@@ -1,15 +1,12 @@
 // wrenfold symbolic code generator.
 // Copyright (c) 2024 Gareth Cross
 // For license information refer to accompanying LICENSE file.
-#define PYBIND11_DETAILED_ERROR_MESSAGES
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>  // Required to pass std::vector.
-
 #include "wf/code_generation/types.h"  //  Required for definition of custom_type.
 #include "wf/compound_expression.h"
 
-#include "docs/compound_expression_wrapper.h"
-#include "wrapper_utils.h"
+#include "pywrenfold/docs/compound_expression_wrapper.h"
+#include "pywrenfold/pybind_imports.h"
+#include "pywrenfold/wrapper_utils.h"
 
 namespace py = pybind11;
 using namespace py::literals;
