+++ components/wrapper/pywrenfold/wrapper_utils.h
@@ -4,11 +4,11 @@
 #pragma once
 #include <algorithm>
 
-#include <pybind11/pybind11.h>
-
 #include "wf/expression.h"
 #include "wf/utility/third_party_imports.h"
 #include "wf/utility/type_list.h"
+
+#include "pywrenfold/pybind_imports.h"
 
 WF_BEGIN_THIRD_PARTY_INCLUDES
 #include <fmt/ranges.h>
@@ -16,7 +16,6 @@ WF_END_THIRD_PARTY_INCLUDES
 
 // Shared utilities for writing wrappers.
 namespace wf {
-namespace py = pybind11;
 
 // Wrap a class, and add certain members automatically.
 template <typename T>
