+++ components/wrapper/pywrenfold/visitor_wrappers.h
@@ -2,12 +2,11 @@
 // Copyright (c) 2024 Gareth Cross
 // For license information refer to accompanying LICENSE file.
 #pragma once
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>  // Required to pass std::vector.
-
 #include "wf/any_expression.h"
 #include "wf/substitute.h"
 #include "wf/utility/algorithms.h"
+
+#include "pywrenfold/pybind_imports.h"
 
 // This file is mostly a gross nuisance to handle the fact that we cannot pass std::variant
 // directly to pybind methods (because it cannot be default initialized). So instead we have to
