+++ components/wrapper/pywrenfold/exceptions_wrapper.cc
@@ -1,12 +1,10 @@
 // wrenfold symbolic code generator.
 // Copyright (c) 2024 Gareth Cross
 // For license information refer to accompanying LICENSE file.
-#define PYBIND11_DETAILED_ERROR_MESSAGES
-#include <pybind11/pybind11.h>
-
 #include "wf/utility/error_types.h"
 
-namespace py = pybind11;
+#include "pywrenfold/pybind_imports.h"
+
 using namespace py::literals;
 
 namespace wf {
