### 2025-01-17 #7303 @ushu
Hi Gareth, @boisgera ,

First, let me apologize for the delayed reply. I’ve now completed my [review checklist](https://github.com/openjournals/joss-reviews/issues/7303#issuecomment-2497908365) to avoid blocking the review process, so everything should proceed smoothly from here.

I want to start by saying that I really enjoyed reviewing your project. The installation process was seamless, and the online documentation (https://wrenfold.org) was clear and helpful. I had the chance to experiment with the provided examples (mostly in C++, since I'm more familiar with it than Rust) and even modify them slightly to explore the functionality further. It feels like a polished, complete product, which is a testament to the amount of work you’ve put into it—well done!

That said, I have a few suggestions for improvement, particularly for the paper. These are listed below in no particular order:

1. **Statement of Need**:
    The statement of need could be expanded by including examples of real-world applications of the framework, particularly in industrial projects. As a competitor to SymForce/SymEngine, the focus on embedded systems is implied, but being explicit about potential use cases would make the paper more compelling for readers.
2. **Usage Example**:
    The “usage example” presented in the paper feels somewhat contrived. It occupies a significant portion of the text but doesn’t add much beyond what the simpler example in Figure 1 already conveys. I’d suggest showcasing more real-world applications where the framework outshines its competitors and limiting the usage example to a concise form (e.g., as shown in Figure 1).
3. **Figure 1** Size:
     In the print version, Figure 1 is quite large (occupying about two-thirds of the page). Reducing its size would improve the overall layout.
4. **Design Rationale**:
    While the tool’s design is briefly outlined, the rationale behind certain decisions could be explained better. For example, why was the symbolic math core implemented in C++ rather than Python? Was this to leverage a specific library (doesn't seem so) or achieve certain performance goals which could not be attained using Python?
5. **Rust Code Generation**:
    One standout feature of Wrenfold is its ability to generate Rust code, which differentiates it from SymForce. Did you have a specific applied project in mind when developing this feature? It would be helpful to discuss the tradeoffs between Rust and C++ and the specific advantages that make Rust a viable choice for certain projects.
6. **Code Integration Example**:
    It might be helpful to include a practical example of how to integrate the generated code with an existing matrix class (using `convert_to_span`). A concise snippet would suffice to illustrate this.
7. **ARM Mac Compatibility**:
    On my M3 Mac, I noticed that no Conda binaries were available, which was surprising. While both pip and build-from-source worked fine, adding Conda binaries would improve the user experience for Mac users.
8. **Reference Formatting**:
    The reference to “Holtz 2022” appears odd in the paper. The author list is overly verbose, including repeated initials, which detracts from readability. Consider revising this.
9. **Future Work**:
    It would be great to include a section outlining planned future developments. For example, do you plan to add new targets (e.g., GPU support) or integrate more closely with other optimization frameworks?

Thank you again for the effort and dedication you’ve put into this project. I’m excited to see how it evolves, and I hope my feedback is helpful in strengthening both the paper and the framework.

Best regards
