### 2025-01-17 #7303 @gareth-cross
Hi @ushu, thank you for your review and the detailed comments. These are great suggestions and I will use them to improve the paper. Regarding point 7 (**ARM Mac Compatibility**), there is a `osx-arm64` [build ](https://anaconda.org/conda-forge/wrenfold) as of version `0.2.2`.

@boisgera A quick question: If I rebase the review branch (`gareth/joss`) off a more recent `main` as I make these changes, will that interrupt the review process at all?
