### 2025-01-18 #7303 @boisgera
> @boisgera A quick question: If I rebase the review branch (gareth/joss) off a more recent main as I make these changes, will that interrupt the review process at all?

Not at all! Since @ushu and @abougouffa have validated their checklist (🙏 thanks to you two by the way!), I can start the "post-review" stage, in which you can still make the changes you deem necessary before you release the latest version of your software.
