+++ paper/paper.bib
@@ -74,3 +74,10 @@
   year = {2022},
   month = {May}
 }
+
+@misc{pybind11,
+   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
+   year = {2016},
+   note = {https://github.com/pybind/pybind11},
+   title = {pybind11 — Seamless operability between C++11 and Python}
+}
