+++ paper/paper.md
@@ -41,11 +41,18 @@ Symbolic code generation can help address these issues:
 * Symbolic functions can be be easily composed in Python, and reasonably performant[^1] C++ implementations are obtained automatically. The developer time cost required to experiment with different optimization parameterizations is thereby reduced.
 * Correct derivatives require no additional work - they can be obtained directly from the objective function via symbolic differentiation. Common terms that appear in the objective function and its Jacobians are automatically de-duplicated by the code generation step.
 
-Compilable source code is a suitable output format because it is straightforward to integrate into downstream projects. By customizing the code generation step, any number of additional languages and development environments can be targeted.
+Compilable source code is a suitable output format because it is straightforward to integrate into downstream projects. By customizing the code generation step, any number of additional languages and development environments can be targeted. Generated functions can be applied to a number of use cases, such as:
+
+* Defining measurement expressions for a factor graph or Kalman filter.[^2]
+* Expressing terms in a motion planning or control optimization.[^3]
 
 [^1]: A comparison with handwritten and auto-diff based implementations is accessible at [https://wrenfold.org/performance.html](https://wrenfold.org/performance.html).
 
-Symbolic code generation has been shown to be an effective tool in robotics. For example, the `MATLAB` symbolic code generation toolbox has been applied directly to motion planning [@Hereid2017FROST]. The open-source `SymForce` framework [@Martiros-RSS-22]  couples the `SymEngine` [@symengine] mathematical backend with Python code generation utilities and mathematical primitives specific to robotics.[^2] `wrenfold` draws inspiration from the design of `SymForce`, while aiming to support a greater variety and complexity of functions. We improve on the concept with the following contributions:
+[^2]: Examples of integrating `wrenfold` functions into Ceres and GTSAM are provided at [https://github.com/wrenfold/wrenfold-extra-examples](https://github.com/wrenfold/wrenfold-extra-examples).
+
+[^3]: A toy example of a Model Predictive Control (MPC) algorithm that uses `wrenfold` is available at [https://github.com/gareth-cross/cart-pole-mpc](https://github.com/gareth-cross/cart-pole-mpc).
+
+Symbolic code generation has been shown to be an effective tool in robotics. For example, the `MATLAB` symbolic code generation toolbox has been applied directly to motion planning [@Hereid2017FROST]. The open-source `SymForce` framework [@Martiros-RSS-22]  couples the `SymEngine` [@symengine] mathematical backend with Python code generation utilities and mathematical primitives specific to robotics.[^4] `wrenfold` draws inspiration from the design of `SymForce`, while aiming to support a greater variety and complexity of functions. We improve on the concept with the following contributions:
 
 * **Symbolic functions may incorporate piecewise conditional statements** - these produce if-else logic in the resulting code. This enables a broader range of functions to be generated.
 * **Emphasis is placed on ease of adaptability of the generated code.** Math expressions are simplified and converted into an abstract syntax tree (AST). Formatting of any element of the AST (such as function signatures and types) can be individually customized by defining a short Python method.
@@ -53,7 +60,7 @@ Symbolic code generation has been shown to be an effective tool in robotics. For
 
 `wrenfold` aims to support researchers and engineers in robotics by bringing symbolic code generation to a greater variety and complexity of optimization problems.
 
-[^2]: `SymForce` has been deployed on production robots produced by Skydio, an American drone manufacturer [@Martiros_Blog_2022].
+[^4]: `SymForce` has been deployed on production robots produced by Skydio, an American drone manufacturer [@Martiros_Blog_2022].
 
 # Design
 
[…]
