### 2025-01-23 #7303 @gareth-cross
Hi @boisgera, I have made a few tweaks to the paper and am ready for your final pass.

Author tasks:
- [x] Make a release of the software with the latest changes from the review and post the version number here: `v0.2.2` [Link to release](https://github.com/wrenfold/wrenfold/releases/tag/v0.2.2)
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. DOI: [10.5281/zenodo.14723141](https://zenodo.org/records/14723141)
