### 2025-01-24 #7303 @boisgera
Hi @gareth-cross,

Could you also change the Zenodo archive title so that it matches the full JOSS article title ("wrenfold: Symbolic code generation for robotics")?
