### 2025-01-25 279bde0
Add DOI for 3 references
