+++ paper/paper.bib
@@ -35,7 +35,8 @@
   address = {Vancouver, BC, Canada},
   month = sep,
   organization = {IEEE/RSJ},
-  month_numeric = {9}
+  month_numeric = {9},
+  doi = {10.1109/iros.2017.8202230}
 }
 
 @book{Barfoot,
@@ -44,7 +45,8 @@
   publisher = {Cambridge University Press},
   year = {2024},
   month = Feb,
-  isbn = {9781009299893}
+  isbn = {9781009299893},
+  doi = {10.1017/9781316671528}
 }
 
 @book{Lynch_Park,
@@ -52,7 +54,8 @@
     publisher = {Cambridge University Press},
     author = {Lynch, Kevin and Park, Frank C.},
     year = {2021},
-    isbn = {9781107156302}
+    isbn = {9781107156302},
+    doi = {10.1017/9781316661239}
 }
 
 @software{symengine,
