### 2025-01-25 #7303 @gareth-cross
> Hi [@gareth-cross](https://github.com/gareth-cross),
>
> Could you have a look at the following DOIs and add them to your bibliography if they check out?
>
> * The suggested [10.1109/iros.2017.8202230](https://doi.org/10.1109/iros.2017.8202230) seems indeed to be the proper DOI for the FROST paper.
> * "State estimation for Robotics"  : [10.1017/9781316671528](https://doi.org/10.1017/9781316671528).
> * "Modern Robotics, Mechanics, Planning, and Control" : [10.1017/9781316661239](https://doi.org/10.1017/9781316661239)

Done, thank you for those suggestions.

> Could you also change the Zenodo archive title so that it matches the full JOSS article title ("wrenfold: Symbolic code generation for robotics")?

Done.
