### 2025-01-28 #7303 @crvernon
:wave: @gareth-cross - I just need you to address the following before I move to accept this one for publication:

In your paper:

- [ ] You are missing the header "References" which should appear at the end of your paper so that when the bib file is compiled it will be inserted after this header.
- [ ] The recommended citation that SymForce provided is not correct.  To ensure the lead author is listed first and the citation is formatted correctly, place commas after each last name.  The same issue occurs with the pybind reference.
- [ ] Also in the pybind reference, please maintain capitalization formatting where it is needed.  You can do this by placing and extra set of curly brackets around the characters you wish to maintain formatting of in the bib file.

Let me know when these changes have been made.  Thanks.
