+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{Martiros-RSS-22,
-    author    = {Hayk Martiros AND Aaron Miller AND Nathan Bucki AND Bradley Solliday AND Ryan Kennedy AND Jack Zhu AND Tung Dang AND Dominic Pattison AND Harrison Zheng AND Teo Tomic AND Peter Henry AND Gareth Cross AND Josiah VanderMey AND Alvin Sun AND Samuel Wang AND Kristen Holtz},
+    author    = {Martiros, Hayk and {{The Skydio Autonomy Team}}},
     title     = {{SymForce: Symbolic Computation and Code Generation for Robotics}},
     booktitle = {Proceedings of Robotics: Science and Systems},
     year      = {2022},
@@ -7,7 +7,7 @@
 }
 
 @software{Agarwal_Ceres_Solver_2022,
-  author = {Agarwal, Sameer and Mierle, Keir and The Ceres Solver Team},
+  author = {Agarwal, Sameer and Mierle, Keir and {{The Ceres Solver Team}}},
   title = {{Ceres Solver}},
   license = {Apache-2.0},
   url = {https://github.com/ceres-solver/ceres-solver},
@@ -17,7 +17,7 @@
 }
 
 @software{gtsam,
-  author       = {Frank Dellaert and GTSAM Contributors},
+  author       = {Dellaert, Frank and {{GTSAM Contributors}}},
   title        = {borglab/gtsam},
   month        = May,
   year         = 2022,
@@ -28,7 +28,7 @@
 }
 
 @inproceedings{Hereid2017FROST,
-  title = {FROST: Fast Robot Optimization and Simulation Toolkit},
+  title = {{FROST: Fast Robot Optimization and Simulation Toolkit}},
   author = {Hereid, Ayonga and Ames, Aaron D.},
   booktitle = {IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   year = {2017},
@@ -40,7 +40,7 @@
 }
 
 @book{Barfoot,
-  title = {State estimation for Robotics},
+  title = {{State estimation for Robotics}},
   author = {Barfoot, Timothy D.},
   publisher = {Cambridge University Press},
   year = {2024},
@@ -50,7 +50,7 @@
 }
 
 @book{Lynch_Park,
-    title = {Modern Robotics: Mechanics, planning, and Control},
+    title = {{Modern Robotics: Mechanics, planning, and Control}},
     publisher = {Cambridge University Press},
     author = {Lynch, Kevin and Park, Frank C.},
     year = {2021},
@@ -59,7 +59,7 @@
 }
 
 @software{symengine,
-	title = {SymEngine},
+	title = {{SymEngine}},
 	url = {https://github.com/symengine/symengine},
 	author = {Fernando, Isuru et. al},
 	date = {2024-06-19},
@@ -69,7 +69,7 @@
 }
 
 @misc{Martiros_Blog_2022,
-  title = {Open-sourcing SymForce},
+  title = {{Open-sourcing SymForce}},
   url = {https://www.skydio.com/blog/open-sourcing-symforce/},
   journal = {Skydio Blog RSS},
   publisher = {Skydio},
@@ -79,8 +79,8 @@
 }
 
 @misc{pybind11,
-   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
+   author = {Jakob, Wenzel and Rhinelander, Jason and Moldovan, Dean},
    year = {2016},
    note = {https://github.com/pybind/pybind11},
-   title = {pybind11 — Seamless operability between C++11 and Python}
+   title = {{pybind11 — Seamless operability between C++11 and Python}}
 }
