### 2025-01-29 #7303 @gareth-cross
Thank you to all involved for your time and effort, I appreciate it. @boisgera @ushu @abougouffa @crvernon

> Please consider becoming a reviewer for JOSS if you are not already: https://reviewers.joss.theoj.org/join

Done!
